saving update to top part spacing and built
