+++ shift-scan/public/sw.js
@@ -1,2 +1,2 @@
-if(!self.define){let e,a={};const i=(i,s)=>(i=new URL(i+".js",s).href,a[i]||new Promise((a=>{if("document"in self){const e=document.createElement("script");e.src=i,e.onload=a,document.head.appendChild(e)}else e=i,importScripts(i),a()})).then((()=>{let e=a[i];if(!e)throw new Error(`Module ${i} didn’t register its module`);return e})));self.define=(s,c)=>{const r=e||("document"in self?document.currentScript.src:"")||location.href;if(a[r])return;let n={};const t=e=>i(e,r),d={module:{uri:r},exports:n,require:t};a[r]=Promise.all(s.map((e=>d[e]||t(e)))).then((e=>(c(...e),n)))}}define(["./workbox-31da5539"],(function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/Settings.svg",revision:"ebe8b6a8bff9bd2fe8669ad036e8edf8"},{url:"/_next/static/PaJipPGFy6ZlwRRVxiksg/_buildManifest.js",revision:"bc2675d22d82f5f5bd07c2ba930f02b2"},{url:"/_next/static/PaJipPGFy6ZlwRRVxiksg/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/chunks/1095-f78c68c0229f4b99.js",revision:"f78c68c0229f4b99"},{url:"/_next/static/chunks/1317-5bb21e343418ec0f.js",revision:"5bb21e343418ec0f"},{url:"/_next/static/chunks/1401-200e2839e2bf3cec.js",revision:"200e2839e2bf3cec"},{url:"/_next/static/chunks/1692-cc16fa14e0e5bbf3.js",revision:"cc16fa14e0e5bbf3"},{url:"/_next/static/chunks/2019-244f4105c4691c0c.js",revision:"244f4105c4691c0c"},{url:"/_next/static/chunks/2057-668deaa0eb77c6a4.js",revision:"668deaa0eb77c6a4"},{url:"/_next/static/chunks/2093-162efb4b28b6a687.js",revision:"162efb4b28b6a687"},{url:"/_next/static/chunks/2170a4aa-8c96009f60e20411.js",revision:"8c96009f60e20411"},{url:"/_next/static/chunks/2328-d7f9841655885b91.js",revision:"d7f9841655885b91"},{url:"/_next/static/chunks/2403.ca107014ddba66de.js",revision:"ca107014ddba66de"},{url:"/_next/static/chunks/2441-9ad2299f689922c4.js",revision:"9ad2299f689922c4"},{url:"/_next/static/chunks/2443-261193c0aabf52ad.js",revision:"261193c0aabf52ad"},{url:"/_next/static/chunks/2541-b9deac5a6c87ddb9.js",revision:"b9deac5a6c87ddb9"},{url:"/_next/static/chunks/2551-2b96b83195377356.js",revision:"2b96b83195377356"},{url:"/_next/static/chunks/2722-48cde404f78497eb.js",revision:"48cde404f78497eb"},{url:"/_next/static/chunks/3068-14a02e90d270dafd.js",revision:"14a02e90d270dafd"},{url:"/_next/static/chunks/3206-67be7e0a1bd37815.js",revision:"67be7e0a1bd37815"},{url:"/_next/static/chunks/3328-1e73249e263f1cb2.js",revision:"1e73249e263f1cb2"},{url:"/_next/static/chunks/3713-01fa541b1020dd9d.js",revision:"01fa541b1020dd9d"},{url:"/_next/static/chunks/3782-7f7d9d67197b4aca.js",revision:"7f7d9d67197b4aca"},{url:"/_next/static/chunks/4095-8d931f3f0218645a.js",revision:"8d931f3f0218645a"},{url:"/_next/static/chunks/4114-b838e9dd1d2a7f67.js",revision:"b838e9dd1d2a7f67"},{url:"/_next/static/chunks/4493-4e6be3f951868868.js",revision:"4e6be3f951868868"},{url:"/_next/static/chunks/475-f9aeee65e8df3294.js",revision:"f9aeee65e8df3294"},{url:"/_next/static/chunks/4914-8e3585d0f1c9f02d.js",revision:"8e3585d0f1c9f02d"},{url:"/_next/static/chunks/4bd1b696-88e37c56edce0f59.js",revision:"88e37c56edce0f59"},{url:"/_next/static/chunks/5152-9e6b4080454ffa3c.js",revision:"9e6b4080454ffa3c"},{url:"/_next/static/chunks/52774a7f-9f526229934f791c.js",revision:"9f526229934f791c"},{url:"/_next/static/chunks/5395-71cc1bac006bc080.js",revision:"71cc1bac006bc080"},{url:"/_next/static/chunks/5493-002a49a6d796cca8.js",revision:"002a49a6d796cca8"},{url:"/_next/static/chunks/5535-ed37b1f77860aa26.js",revision:"ed37b1f77860aa26"},{url:"/_next/static/chunks/5705-2d0aee81a02a655e.js",revision:"2d0aee81a02a655e"},{url:"/_next/static/chunks/5751-277f95667565cf9f.js",revision:"277f95667565cf9f"},{url:"/_next/static/chunks/6323-d1e244e417681874.js",revision:"d1e244e417681874"},{url:"/_next/static/chunks/6544-386355f8b296dc20.js",revision:"386355f8b296dc20"},{url:"/_next/static/chunks/6612-126530b7baba5f69.js",revision:"126530b7baba5f69"},{url:"/_next/static/chunks/6671-c6ce7faab2868fe3.js",revision:"c6ce7faab2868fe3"},{url:"/_next/static/chunks/6697-68520a497e02947b.js",revision:"68520a497e02947b"},{url:"/_next/static/chunks/6862-81cfc3fff123ab8a.js",revision:"81cfc3fff123ab8a"},{url:"/_next/static/chunks/6874-bf3fa41ffddf2fd7.js",revision:"bf3fa41ffddf2fd7"},{url:"/_next/static/chunks/7111-088fe5dbd51639ef.js",revision:"088fe5dbd51639ef"},{url:"/_next/static/chunks/7253-b6a7c4b8ba2a4ccc.js",revision:"b6a7c4b8ba2a4ccc"},{url:"/_next/static/chunks/7655-e651b79933d954e7.js",revision:"e651b79933d954e7"},{url:"/_next/static/chunks/7686-ba6a1435558ab5a4.js",revision:"ba6a1435558ab5a4"},{url:"/_next/static/chunks/789-86a3b92415cd819c.js",revision:"86a3b92415cd819c"},{url:"/_next/static/chunks/822-0edc64ee089dcd0d.js",revision:"0edc64ee089dcd0d"},{url:"/_next/static/chunks/8310-1a53d637c8ce85b0.js",revision:"1a53d637c8ce85b0"},{url:"/_next/static/chunks/8370-4fedd68255eb66d3.js",revision:"4fedd68255eb66d3"},{url:"/_next/static/chunks/8436.fe8b1513f8e56bfa.js",revision:"fe8b1513f8e56bfa"},{url:"/_next/static/chunks/8743.e58f7147681da72a.js",revision:"e58f7147681da72a"},{url:"/_next/static/chunks/9127-f0ee2a66d843196b.js",revision:"f0ee2a66d843196b"},{url:"/_next/static/chunks/9313-eae8d054304ecf16.js",revision:"eae8d054304ecf16"},{url:"/_next/static/chunks/9373-12ddcc70f6722621.js",revision:"12ddcc70f6722621"},{url:"/_next/static/chunks/9481-4cf6b57b1f1291d3.js",revision:"4cf6b57b1f1291d3"},{url:"/_next/static/chunks/9648-8dba046886ed5b99.js",revision:"8dba046886ed5b99"},{url:"/_next/static/chunks/9763-3b7d58f5c287c919.js",revision:"3b7d58f5c287c919"},{url:"/_next/static/chunks/9872-7c697f701ea3a2f2.js",revision:"7c697f701ea3a2f2"},{url:"/_next/static/chunks/app/(routes)/admins/clients/page-47cc4fbfa5d49bab.js",revision:"47cc4fbfa5d49bab"},{url:"/_next/static/chunks/app/(routes)/admins/cost-codes/page-330bf2b6bdfeb166.js",revision:"330bf2b6bdfeb166"},{url:"/_next/static/chunks/app/(routes)/admins/crew/page-97e2b1238b28e535.js",revision:"97e2b1238b28e535"},{url:"/_next/static/chunks/app/(routes)/admins/equipment/page-297718791ffac793.js",revision:"297718791ffac793"},{url:"/_next/static/chunks/app/(routes)/admins/forms/%5Bid%5D/page-e1d2eb047463bac8.js",revision:"e1d2eb047463bac8"},{url:"/_next/static/chunks/app/(routes)/admins/forms/create/page-5efe530cd05cb6db.js",revision:"5efe530cd05cb6db"},{url:"/_next/static/chunks/app/(routes)/admins/forms/edit/%5Bid%5D/page-c6b5488e0a16dce1.js",revision:"c6b5488e0a16dce1"},{url:"/_next/static/chunks/app/(routes)/admins/forms/page-6a37dfa3a833a8bf.js",revision:"6a37dfa3a833a8bf"},{url:"/_next/static/chunks/app/(routes)/admins/jobsites/page-70c68ef28253f118.js",revision:"70c68ef28253f118"},{url:"/_next/static/chunks/app/(routes)/admins/layout-74b03a55699d8958.js",revision:"74b03a55699d8958"},{url:"/_next/static/chunks/app/(routes)/admins/page-cdc3b85b0c3243d3.js",revision:"cdc3b85b0c3243d3"},{url:"/_next/static/chunks/app/(routes)/admins/personnel/page-e30342af2b941d29.js",revision:"e30342af2b941d29"},{url:"/_next/static/chunks/app/(routes)/admins/reports/page-309b6d6b4f3e486f.js",revision:"309b6d6b4f3e486f"},{url:"/_next/static/chunks/app/(routes)/admins/timesheets/page-f46894989290ffe4.js",revision:"f46894989290ffe4"},{url:"/_next/static/chunks/app/(routes)/break/page-21caaf8314224166.js",revision:"21caaf8314224166"},{url:"/_next/static/chunks/app/(routes)/clock/page-916e93cebf243368.js",revision:"916e93cebf243368"},{url:"/_next/static/chunks/app/(routes)/dashboard/clock-out/(components)/log/page-b6e632fe83e2498b.js",revision:"b6e632fe83e2498b"},{url:"/_next/static/chunks/app/(routes)/dashboard/clock-out/page-df61dcb16cfc00f3.js",revision:"df61dcb16cfc00f3"},{url:"/_next/static/chunks/app/(routes)/dashboard/equipment/%5Bid%5D/page-24ccbf9235edbe67.js",revision:"24ccbf9235edbe67"},{url:"/_next/static/chunks/app/(routes)/dashboard/equipment/log-new/page-0e751569e3f67681.js",revision:"0e751569e3f67681"},{url:"/_next/static/chunks/app/(routes)/dashboard/equipment/page-01c72909cb40becf.js",revision:"01c72909cb40becf"},{url:"/_next/static/chunks/app/(routes)/dashboard/mechanic/edit-repair-details/%5Bid%5D/page-56d0aeae424b938f.js",revision:"56d0aeae424b938f"},{url:"/_next/static/chunks/app/(routes)/dashboard/mechanic/new-repair/page-490823836af9207c.js",revision:"490823836af9207c"},{url:"/_next/static/chunks/app/(routes)/dashboard/mechanic/page-341d41682cc8b6a6.js",revision:"341d41682cc8b6a6"},{url:"/_next/static/chunks/app/(routes)/dashboard/mechanic/projects/%5Bid%5D/page-55aeee7523b72689.js",revision:"55aeee7523b72689"},{url:"/_next/static/chunks/app/(routes)/dashboard/myTeam/%5Bid%5D/employee/%5BemployeeId%5D/page-0dda2ae04de3ded2.js",revision:"0dda2ae04de3ded2"},{url:"/_next/static/chunks/app/(routes)/dashboard/myTeam/%5Bid%5D/page-ccb5d2e7958313d7.js",revision:"ccb5d2e7958313d7"},{url:"/_next/static/chunks/app/(routes)/dashboard/myTeam/page-73437efe4ff35b75.js",revision:"73437efe4ff35b75"},{url:"/_next/static/chunks/app/(routes)/dashboard/myTeam/timecards/page-9eb9b56fc9ccb487.js",revision:"9eb9b56fc9ccb487"},{url:"/_next/static/chunks/app/(routes)/dashboard/page-4016bf3e76d83acb.js",revision:"4016bf3e76d83acb"},{url:"/_next/static/chunks/app/(routes)/dashboard/qr-generator/add-equipment/page-92380df793993c51.js",revision:"92380df793993c51"},{url:"/_next/static/chunks/app/(routes)/dashboard/qr-generator/add-jobsite/page-bf715c823804df6f.js",revision:"bf715c823804df6f"},{url:"/_next/static/chunks/app/(routes)/dashboard/qr-generator/page-76d6385b787361da.js",revision:"76d6385b787361da"},{url:"/_next/static/chunks/app/(routes)/dashboard/switch-jobs/page-c6da0ba6d1e256f0.js",revision:"c6da0ba6d1e256f0"},{url:"/_next/static/chunks/app/(routes)/dashboard/tasco/page-8a3f0c2b107e2d45.js",revision:"8a3f0c2b107e2d45"},{url:"/_next/static/chunks/app/(routes)/dashboard/truckingAssistant/page-2ccb8d75e7fb553e.js",revision:"2ccb8d75e7fb553e"},{url:"/_next/static/chunks/app/(routes)/hamburger/changePassword/page-38d824dd29b7f69c.js",revision:"38d824dd29b7f69c"},{url:"/_next/static/chunks/app/(routes)/hamburger/inbox/formSubmission/%5Bid%5D/page-50a86d5272819b5b.js",revision:"50a86d5272819b5b"},{url:"/_next/static/chunks/app/(routes)/hamburger/inbox/page-123a92a845220059.js",revision:"123a92a845220059"},{url:"/_next/static/chunks/app/(routes)/hamburger/profile/page-70cbae58e9e16cdb.js",revision:"70cbae58e9e16cdb"},{url:"/_next/static/chunks/app/(routes)/information/page-89d7583d61422b6d.js",revision:"89d7583d61422b6d"},{url:"/_next/static/chunks/app/(routes)/signin/forgot-password/page-452ac6e4dbf00e2c.js",revision:"452ac6e4dbf00e2c"},{url:"/_next/static/chunks/app/(routes)/signin/new-password/page-48db32da7718445f.js",revision:"48db32da7718445f"},{url:"/_next/static/chunks/app/(routes)/signin/page-776eda5d900bfbb8.js",revision:"776eda5d900bfbb8"},{url:"/_next/static/chunks/app/(routes)/signin/signup/page-fd3ccee82d025100.js",revision:"fd3ccee82d025100"},{url:"/_next/static/chunks/app/(routes)/timesheets/page-770ac7911f8bc820.js",revision:"770ac7911f8bc820"},{url:"/_next/static/chunks/app/_not-found/page-e32296c53d3e4077.js",revision:"e32296c53d3e4077"},{url:"/_next/static/chunks/app/api/(admins)/crewManager/route-a6ad4fd50bc07e88.js",revision:"a6ad4fd50bc07e88"},{url:"/_next/static/chunks/app/api/(admins)/employeeInfo/%5Bemployee%5D/route-b8fc451918d6fb9e.js",revision:"b8fc451918d6fb9e"},{url:"/_next/static/chunks/app/api/(admins)/getAllActiveEmployeeName/route-ad31be4f725c1a32.js",revision:"ad31be4f725c1a32"},{url:"/_next/static/chunks/app/api/(admins)/getAllCostCodesByJobSites/route-0013863ff3a6e8dc.js",revision:"0013863ff3a6e8dc"},{url:"/_next/static/chunks/app/api/(admins)/getAllCrews/route-8a68e32a18b4e2f1.js",revision:"8a68e32a18b4e2f1"},{url:"/_next/static/chunks/app/api/(admins)/getAllEmployees/route-96c691df49827c48.js",revision:"96c691df49827c48"},{url:"/_next/static/chunks/app/api/(admins)/getAllEquipment/route-47258141a670c4a4.js",revision:"47258141a670c4a4"},{url:"/_next/static/chunks/app/api/(admins)/getAllForms/route-5ed55ec535bd3c90.js",revision:"5ed55ec535bd3c90"},{url:"/_next/static/chunks/app/api/(admins)/getAllTimesheetInfo/route-3b13943fe248499b.js",revision:"3b13943fe248499b"},{url:"/_next/static/chunks/app/api/(admins)/getAllTimesheetsPending/route-bdae5e866282ceef.js",revision:"bdae5e866282ceef"},{url:"/_next/static/chunks/app/api/(admins)/getClientById/%5Bid%5D/route-c2cccbabe90740c9.js",revision:"c2cccbabe90740c9"},{url:"/_next/static/chunks/app/api/(admins)/getClientsDetails/route-aaca05f497bbcc55.js",revision:"aaca05f497bbcc55"},{url:"/_next/static/chunks/app/api/(admins)/getCostCodeDetails/route-813fb26c1b04b60f.js",revision:"813fb26c1b04b60f"},{url:"/_next/static/chunks/app/api/(admins)/getCostCodeSummary/route-90bd8f9764c1e92e.js",revision:"90bd8f9764c1e92e"},{url:"/_next/static/chunks/app/api/(admins)/getCostCodeTags/%5Bid%5D/route-f1aa8ee55e03ce8e.js",revision:"f1aa8ee55e03ce8e"},{url:"/_next/static/chunks/app/api/(admins)/getCrewByIdAdmin/%5Bid%5D/route-6e363c5f3aef3c06.js",revision:"6e363c5f3aef3c06"},{url:"/_next/static/chunks/app/api/(admins)/getDashboard/route-be74688f5dce0471.js",revision:"be74688f5dce0471"},{url:"/_next/static/chunks/app/api/(admins)/getEquipmentByEquipmentId/%5Bequipment%5D/route-f843dc2e35581a19.js",revision:"f843dc2e35581a19"},{url:"/_next/static/chunks/app/api/(admins)/getEquipmentDetails/route-aed522d0f2ca28f9.js",revision:"aed522d0f2ca28f9"},{url:"/_next/static/chunks/app/api/(admins)/getEquipmentSummary/route-e9b56c06596eedaf.js",revision:"e9b56c06596eedaf"},{url:"/_next/static/chunks/app/api/(admins)/getFormSubmissionsById/%5Bid%5D/route-0b3a0cd9de35ee70.js",revision:"0b3a0cd9de35ee70"},{url:"/_next/static/chunks/app/api/(admins)/getForms/%5Bid%5D/route-82ece0b3426108d2.js",revision:"82ece0b3426108d2"},{url:"/_next/static/chunks/app/api/(admins)/getJobsiteById/%5Bid%5D/route-a4ca1a73b8757ec1.js",revision:"a4ca1a73b8757ec1"},{url:"/_next/static/chunks/app/api/(admins)/getJobsiteSummary/route-65701a5a0abb06ad.js",revision:"65701a5a0abb06ad"},{url:"/_next/static/chunks/app/api/(admins)/getMechanicProjectSummary/route-c843295f63c4e44f.js",revision:"c843295f63c4e44f"},{url:"/_next/static/chunks/app/api/(admins)/getTagById/%5Bid%5D/route-3a8d0f9ef7cb2ced.js",revision:"3a8d0f9ef7cb2ced"},{url:"/_next/static/chunks/app/api/(admins)/getTagSummary/route-43391b409c40c428.js",revision:"43391b409c40c428"},{url:"/_next/static/chunks/app/api/(admins)/getTimesheetById/route-e8b48214da8d3a54.js",revision:"e8b48214da8d3a54"},{url:"/_next/static/chunks/app/api/(admins)/jobsiteManager/route-9a4b336f15428190.js",revision:"9a4b336f15428190"},{url:"/_next/static/chunks/app/api/(admins)/personnelManager/route-02f9e90102c4e564.js",revision:"02f9e90102c4e564"},{url:"/_next/static/chunks/app/api/(admins)/reports/tasco/route-46a2f5328cf22bf1.js",revision:"46a2f5328cf22bf1"},{url:"/_next/static/chunks/app/api/(admins)/reports/truckingReport/route-4a88ee9b71036746.js",revision:"4a88ee9b71036746"},{url:"/_next/static/chunks/app/api/(clock)/getEquipmentList/route-4730f2eef80c4bcc.js",revision:"4730f2eef80c4bcc"},{url:"/_next/static/chunks/app/api/(clock)/getLogs/route-41e2b14cc47b8324.js",revision:"41e2b14cc47b8324"},{url:"/_next/static/chunks/app/api/(clock)/getMyTeamsUsers/route-1de587d2af723a09.js",revision:"1de587d2af723a09"},{url:"/_next/static/chunks/app/api/(clock)/getPendingTeamTimesheets/%5BcrewMembers%5D/route-030c108b1524f10d.js",revision:"030c108b1524f10d"},{url:"/_next/static/chunks/app/api/(clock)/getRecentCostCodes/route-66011e59ba798626.js",revision:"66011e59ba798626"},{url:"/_next/static/chunks/app/api/(clock)/getTodaysTimesheets/route-a7014ca9e8477ab2.js",revision:"a7014ca9e8477ab2"},{url:"/_next/static/chunks/app/api/(clock)/getTruckData/route-d43e55d8abea453d.js",revision:"d43e55d8abea453d"},{url:"/_next/static/chunks/app/api/(equipment)/getAllEquipmentIdAndQrId/route-c8f9906ad1ec689e.js",revision:"c8f9906ad1ec689e"},{url:"/_next/static/chunks/app/api/(equipment)/getCheckedList/route-6aa444051fc09b44.js",revision:"6aa444051fc09b44"},{url:"/_next/static/chunks/app/api/(equipment)/getEqUserLogs/%5Bid%5D/route-51627412ebb068bc.js",revision:"51627412ebb068bc"},{url:"/_next/static/chunks/app/api/(equipment)/getEqUserLogs/route-e9010271cbf63852.js",revision:"e9010271cbf63852"},{url:"/_next/static/chunks/app/api/(equipment)/getEquipmentRefueledLogs/%5Bid%5D/route-5c254f91ac2260c4.js",revision:"5c254f91ac2260c4"},{url:"/_next/static/chunks/app/api/(equipment)/getRecentJobDetails/route-b51c51dccc4f5d98.js",revision:"b51c51dccc4f5d98"},{url:"/_next/static/chunks/app/api/(forms)/employeeRequests/%5Bfilter%5D/route-108c4aea7f16e672.js",revision:"108c4aea7f16e672"},{url:"/_next/static/chunks/app/api/(forms)/form/%5Bid%5D/route-def5d21120e4c1c6.js",revision:"def5d21120e4c1c6"},{url:"/_next/static/chunks/app/api/(forms)/formDraft/%5Bid%5D/route-c27f4097ff3cec17.js",revision:"c27f4097ff3cec17"},{url:"/_next/static/chunks/app/api/(forms)/formDrafts/route-0c1896c75aa610df.js",revision:"0c1896c75aa610df"},{url:"/_next/static/chunks/app/api/(forms)/formSubmission/%5Bid%5D/route-4a77a032c874ef51.js",revision:"4a77a032c874ef51"},{url:"/_next/static/chunks/app/api/(forms)/formSubmissions/%5Bstatus%5D/route-2b172217bfd31857.js",revision:"2b172217bfd31857"},{url:"/_next/static/chunks/app/api/(forms)/forms/route-386e9b6eadce202b.js",revision:"386e9b6eadce202b"},{url:"/_next/static/chunks/app/api/(forms)/getDocumentById/%5Bid%5D/route-16a72506935bb68b.js",revision:"16a72506935bb68b"},{url:"/_next/static/chunks/app/api/(forms)/getDocumentTags/route-0d1d2efd90b7d4a0.js",revision:"0d1d2efd90b7d4a0"},{url:"/_next/static/chunks/app/api/(forms)/getDocuments/route-5735936db823474e.js",revision:"5735936db823474e"},{url:"/_next/static/chunks/app/api/(forms)/getEmployees/route-85b42aa19cf4b68e.js",revision:"85b42aa19cf4b68e"},{url:"/_next/static/chunks/app/api/(forms)/getEqForDocs/route-71e475a4876e7ff9.js",revision:"71e475a4876e7ff9"},{url:"/_next/static/chunks/app/api/(forms)/managerFormApproval/%5Bid%5D/route-4ad2749ba3cb8a8e.js",revision:"4ad2749ba3cb8a8e"},{url:"/_next/static/chunks/app/api/(forms)/teamSubmission/%5Bid%5D/route-e2fcfe03875ad2de.js",revision:"e2fcfe03875ad2de"},{url:"/_next/static/chunks/app/api/(general)/cookies/route-f198ae2c75bb0add.js",revision:"f198ae2c75bb0add"},{url:"/_next/static/chunks/app/api/(general)/getAllJobsites/route-4f2e0626ab8c78d2.js",revision:"4f2e0626ab8c78d2"},{url:"/_next/static/chunks/app/api/(general)/getAllTags/route-a8ba9526a1d939dd.js",revision:"a8ba9526a1d939dd"},{url:"/_next/static/chunks/app/api/(general)/getBannerData/route-ce96a4f79060d576.js",revision:"ce96a4f79060d576"},{url:"/_next/static/chunks/app/api/(general)/getCostCodes/route-839dc537fcbbd0ae.js",revision:"839dc537fcbbd0ae"},{url:"/_next/static/chunks/app/api/(general)/getCrewByCrewId/%5Bcrew%5D/route-04816e0d35252339.js",revision:"04816e0d35252339"},{url:"/_next/static/chunks/app/api/(general)/getEquipment/route-6f063fe970202a3c.js",revision:"6f063fe970202a3c"},{url:"/_next/static/chunks/app/api/(general)/getPayPeriodTimeSheets/route-ee7337de72860ecc.js",revision:"ee7337de72860ecc"},{url:"/_next/static/chunks/app/api/(general)/getRecentEquipment/route-bda9e71b194a8724.js",revision:"bda9e71b194a8724"},{url:"/_next/static/chunks/app/api/(general)/getRecentJobsites/route-5fc7e4a03a80fef1.js",revision:"5fc7e4a03a80fef1"},{url:"/_next/static/chunks/app/api/(general)/getRecentTimecard/route-63c4a5bb3607cfba.js",revision:"63c4a5bb3607cfba"},{url:"/_next/static/chunks/app/api/(general)/getRecentTimecardReturn/route-c4a891e6b10397d8.js",revision:"c4a891e6b10397d8"},{url:"/_next/static/chunks/app/api/(general)/getSettings/route-25a1789704590a17.js",revision:"25a1789704590a17"},{url:"/_next/static/chunks/app/api/(general)/getTimesheets/%5Bemployee%5D/route-d19db3e142dc44e4.js",revision:"d19db3e142dc44e4"},{url:"/_next/static/chunks/app/api/(general)/getTimesheets/route-e038697331997d18.js",revision:"e038697331997d18"},{url:"/_next/static/chunks/app/api/(general)/getUserSignature/route-51aaa98dc7282071.js",revision:"51aaa98dc7282071"},{url:"/_next/static/chunks/app/api/(mechanic)/getMaintenanceLogs/%5Bid%5D/route-bd938afe60677879.js",revision:"bd938afe60677879"},{url:"/_next/static/chunks/app/api/(mechanic)/getMaintenanceProjects/route-18b286a2359e9dfe.js",revision:"18b286a2359e9dfe"},{url:"/_next/static/chunks/app/api/(mechanic)/getReceivedInfo/%5Bid%5D/route-7544b71bd88555ea.js",revision:"7544b71bd88555ea"},{url:"/_next/static/chunks/app/api/(mechanic)/getRepairDetails/%5Bid%5D/route-089ac4bd0da69a42.js",revision:"089ac4bd0da69a42"},{url:"/_next/static/chunks/app/api/(myTeam)/getCrew/route-9d1d447937f92205.js",revision:"9d1d447937f92205"},{url:"/_next/static/chunks/app/api/(myTeam)/getCrewById/%5BcrewId%5D/route-97af907f88f480fa.js",revision:"97af907f88f480fa"},{url:"/_next/static/chunks/app/api/(myTeam)/getPendingTeamTimeSheets/route-8aa5bcff4e5e0ec6.js",revision:"8aa5bcff4e5e0ec6"},{url:"/_next/static/chunks/app/api/(myTeam)/getTeam/route-1d9bbd33c8160a21.js",revision:"1d9bbd33c8160a21"},{url:"/_next/static/chunks/app/api/(myTeam)/getUserInfo/%5Bid%5D/route-d3d91f6dee197267.js",revision:"d3d91f6dee197267"},{url:"/_next/static/chunks/app/api/(qrGenerator)/getEquipmentbyQrId/%5BqrId%5D/route-ee6b4690eb4b1acb.js",revision:"ee6b4690eb4b1acb"},{url:"/_next/static/chunks/app/api/(qrGenerator)/getJobsites/route-ec72fb07766b4045.js",revision:"ec72fb07766b4045"},{url:"/_next/static/chunks/app/api/(reviewWithDevun)/getSignature/%5Bid%5D/route-690be204793a0258.js",revision:"690be204793a0258"},{url:"/_next/static/chunks/app/api/(reviewWithDevun)/getTimesheetById/%5Bemployee%5D/route-a1410e3a9fd6a75d.js",revision:"a1410e3a9fd6a75d"},{url:"/_next/static/chunks/app/api/(settings)/getEmployee/route-7ac9828e0cf7acb0.js",revision:"7ac9828e0cf7acb0"},{url:"/_next/static/chunks/app/api/(tasco)/getTascoLog/comment/%5BtimeSheetId%5D/route-db2514bbe36d609a.js",revision:"db2514bbe36d609a"},{url:"/_next/static/chunks/app/api/(tasco)/getTascoLog/loadCount/%5BtimeSheetId%5D/route-e1d7e348685bd47e.js",revision:"e1d7e348685bd47e"},{url:"/_next/static/chunks/app/api/(tasco)/getTascoLog/refueledLogs/%5BtimeSheetId%5D/route-048e1a7248d42c53.js",revision:"048e1a7248d42c53"},{url:"/_next/static/chunks/app/api/(tasco)/getTascoLog/tascoId/route-4c82b5f5bdce889d.js",revision:"4c82b5f5bdce889d"},{url:"/_next/static/chunks/app/api/(timeSheets)/getComment/route-000b35c9a7c0e1ec.js",revision:"000b35c9a7c0e1ec"},{url:"/_next/static/chunks/app/api/(timeSheets)/getMaterialTypes/route-f6d6864a813f60fb.js",revision:"f6d6864a813f60fb"},{url:"/_next/static/chunks/app/api/(trucking)/getTruckingLogs/endingMileage/%5BtimeSheetId%5D/route-5eafa7cfe7885ec8.js",revision:"5eafa7cfe7885ec8"},{url:"/_next/static/chunks/app/api/(trucking)/getTruckingLogs/equipmentHauled/%5BtimeSheetId%5D/route-56267620569b4d7f.js",revision:"56267620569b4d7f"},{url:"/_next/static/chunks/app/api/(trucking)/getTruckingLogs/material/%5BtimeSheetId%5D/route-77bd5bff7071b9f1.js",revision:"77bd5bff7071b9f1"},{url:"/_next/static/chunks/app/api/(trucking)/getTruckingLogs/notes/%5BtimeSheetId%5D/route-935cbb96f3d803f5.js",revision:"935cbb96f3d803f5"},{url:"/_next/static/chunks/app/api/(trucking)/getTruckingLogs/refueledLogs/%5BtimeSheetId%5D/route-e3a0e81daba2f047.js",revision:"e3a0e81daba2f047"},{url:"/_next/static/chunks/app/api/(trucking)/getTruckingLogs/startingMileage/%5BtimeSheetId%5D/route-8608c51eb55c9658.js",revision:"8608c51eb55c9658"},{url:"/_next/static/chunks/app/api/(trucking)/getTruckingLogs/stateMileage/%5BtimeSheetId%5D/route-8160ab7228d3e6e1.js",revision:"8160ab7228d3e6e1"},{url:"/_next/static/chunks/app/api/(trucking)/getTruckingLogs/truckingId/route-dd0edb856a340d50.js",revision:"dd0edb856a340d50"},{url:"/_next/static/chunks/app/api/(user)/getUserImage/route-6adf5a7de0093e39.js",revision:"6adf5a7de0093e39"},{url:"/_next/static/chunks/app/api/auth/%5B...nextauth%5D/route-6d1e12b682014910.js",revision:"6d1e12b682014910"},{url:"/_next/static/chunks/app/api/clockoutDetails/route-352ff9e51ca65180.js",revision:"352ff9e51ca65180"},{url:"/_next/static/chunks/app/api/equipment/%5BequipmentId%5D/lastMileage/route-94419db7d73a8237.js",revision:"94419db7d73a8237"},{url:"/_next/static/chunks/app/api/getTimesheetDetailsManager/%5Bid%5D/route-e01aade37f72fe2b.js",revision:"e01aade37f72fe2b"},{url:"/_next/static/chunks/app/api/getTimesheetsByDate/route-d3773ff605e31d2e.js",revision:"d3773ff605e31d2e"},{url:"/_next/static/chunks/app/api/getTimesheetsByDateNew/route-d1db4096be08a38b.js",revision:"d1db4096be08a38b"},{url:"/_next/static/chunks/app/api/getTruckingLogs/equipmentId/%5BtimeSheetId%5D/route-3e1bd301359b6d6a.js",revision:"3e1bd301359b6d6a"},{url:"/_next/static/chunks/app/api/sentry-example-api/route-3ad3b080b6289ca2.js",revision:"3ad3b080b6289ca2"},{url:"/_next/static/chunks/app/api/truckingLogDetails/%5BtruckingLogId%5D/route-f424dae265011067.js",revision:"f424dae265011067"},{url:"/_next/static/chunks/app/api/user-permissions/route-49bef76c98d83ef7.js",revision:"49bef76c98d83ef7"},{url:"/_next/static/chunks/app/error-8ed5a2d151f85834.js",revision:"8ed5a2d151f85834"},{url:"/_next/static/chunks/app/global-error-293467dba0cb08e6.js",revision:"293467dba0cb08e6"},{url:"/_next/static/chunks/app/icons/page-c8cf36ee1169cb53.js",revision:"c8cf36ee1169cb53"},{url:"/_next/static/chunks/app/layout-923910bbb72323dc.js",revision:"923910bbb72323dc"},{url:"/_next/static/chunks/app/not-authorized/page-1f20afeac966c607.js",revision:"1f20afeac966c607"},{url:"/_next/static/chunks/app/not-found-25775581529f7a18.js",revision:"25775581529f7a18"},{url:"/_next/static/chunks/app/page-3d0d29b8136fb1b4.js",revision:"3d0d29b8136fb1b4"},{url:"/_next/static/chunks/framework-5871c7b6f4022a8e.js",revision:"5871c7b6f4022a8e"},{url:"/_next/static/chunks/main-app-511dc06255e2bc08.js",revision:"511dc06255e2bc08"},{url:"/_next/static/chunks/main-f3835dd8dd9ffcc4.js",revision:"f3835dd8dd9ffcc4"},{url:"/_next/static/chunks/pages/_app-5376b67503103dac.js",revision:"5376b67503103dac"},{url:"/_next/static/chunks/pages/_error-e94a0482ccaf0c05.js",revision:"e94a0482ccaf0c05"},{url:"/_next/static/chunks/polyfills-42372ed130431b0a.js",revision:"846118c33b2c0e922d7b3a7676f81f6f"},{url:"/_next/static/chunks/webpack-7508aada89e7c79a.js",revision:"7508aada89e7c79a"},{url:"/_next/static/css/6331649fe4665ef9.css",revision:"6331649fe4665ef9"},{url:"/_next/static/css/63c7182515913e2f.css",revision:"63c7182515913e2f"},{url:"/_next/static/css/ae3f615d9a81d81a.css",revision:"ae3f615d9a81d81a"},{url:"/_next/static/css/b5fe2f5a95132839.css",revision:"b5fe2f5a95132839"},{url:"/_next/static/css/ccce47fdcedd91d2.css",revision:"ccce47fdcedd91d2"},{url:"/_next/static/media/30e6a981a0efeb9b-s.woff2",revision:"5cd14db2bb47ab5db74376b77c909759"},{url:"/_next/static/media/5df40a3eb8df1a4c-s.p.woff2",revision:"2f46a0d75dfff5f623a08402583e7ee9"},{url:"/_next/static/media/62d466ce3199db61-s.p.woff2",revision:"3a28d888908be7f560bf59faa1c796b2"},{url:"/_next/static/media/92c2336771404627-s.woff2",revision:"e456d9e276ba68ad76f65962e88fab81"},{url:"/_next/static/media/ajax-loader.0b80f665.gif",revision:"0b80f665"},{url:"/_next/static/media/d934919622a522a7-s.woff2",revision:"d3322001406183e84b58b99afd278b84"},{url:"/_next/static/media/slick.25572f22.eot",revision:"25572f22"},{url:"/_next/static/media/slick.653a4cbb.woff",revision:"653a4cbb"},{url:"/_next/static/media/slick.6aa1ee46.ttf",revision:"6aa1ee46"},{url:"/_next/static/media/slick.f895cfdf.svg",revision:"f895cfdf"},{url:"/admin-white.svg",revision:"f60bbe82328c01c11b8e0b2970c889e6"},{url:"/admin.svg",revision:"0d220493364872b69d6f6cde6e385d2c"},{url:"/android/android-launchericon-144-144.png",revision:"554cb47f9b447d7c049e0af337f30598"},{url:"/android/android-launchericon-192-192.png",revision:"c7acce3a1ca251dc273817cd5462a666"},{url:"/android/android-launchericon-48-48.png",revision:"b3e487bbda706b0a7a31bd383fcb5bb9"},{url:"/android/android-launchericon-512-512.png",revision:"daaca9c8424139dcfdeed3181b0a7cbd"},{url:"/android/android-launchericon-72-72.png",revision:"0544816fbb562d0825feb6e9ff73edff"},{url:"/android/android-launchericon-96-96.png",revision:"29c67668a313466cbd66948974b9c656"},{url:"/arrowBack.svg",revision:"17d544aac2962f3f0b83d7dd8bf8d5e7"},{url:"/arrowDown.svg",revision:"bde6e8ea1fa48902e773fc0f34d8ad2c"},{url:"/arrowDownThin.svg",revision:"c8dc31e8dcdca0b4feb827ad7b28c336"},{url:"/arrowLeft.svg",revision:"548bd36e974c05627ee6042a0d17e6c6"},{url:"/arrowLeftSymbol.svg",revision:"4cfdc72b6b5f2cf11ea191dc60c5c4b4"},{url:"/arrowLeftThin.svg",revision:"c4bbdff0ab65d617688561fee993f1c1"},{url:"/arrowRight.svg",revision:"36ae3b90062b7432459cad98be741a80"},{url:"/arrowRightSymbol.svg",revision:"e771082a8a75d50f59df162c67d26091"},{url:"/arrowRightThin.svg",revision:"73a5a9d7e4bf4951d11c4e268faeee85"},{url:"/arrowUp.svg",revision:"3e3a91581069f2b5dc59c7c465255263"},{url:"/arrowUpThin.svg",revision:"ef72218a70f19849149176593f9a8cee"},{url:"/biometrics.svg",revision:"f42bc2055fc95544b3641a2819a084d7"},{url:"/broken.svg",revision:"c62a555a912a5887033a4661236ae31d"},{url:"/calendar.svg",revision:"d6aa1ba02d5a6253fcec8da68242da2a"},{url:"/camera.svg",revision:"4657cc3ac5e2a543f0829e6100446591"},{url:"/cameraFilled.svg",revision:"bcdbce7deaeb9a1d37058d3befea92fc"},{url:"/checkbox.svg",revision:"c2c40521ced6fc4ba77a3865d9dc345f"},{url:"/clock.svg",revision:"eff35cd79029bdd0eeebc39b052f0fbf"},{url:"/clockBreak.svg",revision:"0f0d1da56fbbed48b97b5dbba7d34e65"},{url:"/clockIn.svg",revision:"02e58b9c6f4a129a7026400944bd15b7"},{url:"/clockOut.svg",revision:"a8b9f471daf782532b4006c2cb20b89a"},{url:"/comment.svg",revision:"7bb5715a71c021149b5c3b811aac572b"},{url:"/comments.svg",revision:"29b1a7eb27c708878a454c2eeaad616d"},{url:"/condense-white.svg",revision:"6e9075aa9d7fcb1c242b78610ea9330a"},{url:"/condense.svg",revision:"bc4d6b17d44724df96d6d4a3016072da"},{url:"/drag.svg",revision:"da36f043748f4e8af0e0f5ee17e289a2"},{url:"/dragDots.svg",revision:"c9d25bda626917ea72ca49977f11a016"},{url:"/dragFormBuilder.svg",revision:"bfdedbf1bc6d64afb3642848a542f55d"},{url:"/endDay.svg",revision:"78010192db1eaa29e95ac4d2d9f3705e"},{url:"/equipment-white.svg",revision:"0760c1e58029025d443832cdc33d7402"},{url:"/equipment.svg",revision:"eeb87295d51e91e2adb6483c18bc2452"},{url:"/eraser.svg",revision:"7ed7c543a0ff5530d3c586b368c51315"},{url:"/export-white.svg",revision:"4793d42ec50e11a8586bbc2a5a2f59c8"},{url:"/export.svg",revision:"09f9524501c9e871487735c2b09fcc43"},{url:"/eye-blue.svg",revision:"a851e7f5c035279390dfa1eaa3387e0d"},{url:"/eye.svg",revision:"b7243fcb645ec88a90d9b66fdc2abca3"},{url:"/eyeSlash.svg",revision:"3bb6de382b51b3a333e146bcf8a5d4f5"},{url:"/favicon_io/android-chrome-192x192.png",revision:"07281900d7ffa5690cc984da6e613805"},{url:"/favicon_io/android-chrome-512x512.png",revision:"ee520bf09fd4e101b8952887694905fc"},{url:"/favicon_io/apple-touch-icon.png",revision:"6f682a595255e6aab2bf0f2daa78403a"},{url:"/favicon_io/favicon-16x16.png",revision:"2d00862eea3e0a26607d3d0ce24d1747"},{url:"/favicon_io/favicon-32x32.png",revision:"84f7a4563c6fb8cd905f2dbedcffede8"},{url:"/fileClosed.svg",revision:"bf5a8aeebd62c1736f9e0062a3a271a0"},{url:"/fileOpen.svg",revision:"9ce1ceadae4b4abba237029e0a410d8f"},{url:"/filterDials.svg",revision:"13341b5339e076be7b361c2e09e4976c"},{url:"/filterFunnel.svg",revision:"743604a4e2e50abbb135746948880f65"},{url:"/form-white.svg",revision:"5f85274db63c253161739424aa2fab34"},{url:"/form.svg",revision:"5828ab7b08b01307092a48136ac39cb0"},{url:"/formApproval.svg",revision:"a53cba1fffa0cf55e4dbd86d0345b5e8"},{url:"/formDuplicate.svg",revision:"898bd36ec2c706978dae4e684575ec3e"},{url:"/formEdit.svg",revision:"017c59837c434df47dce4b457e71996e"},{url:"/formInspect-white.svg",revision:"8e48a64eb645a8b5bdc204706e8e6451"},{url:"/formInspect.svg",revision:"f2fcd14456afc03a583c568f4c4da3c8"},{url:"/formList-white.svg",revision:"e1ced4954a119850e3f06e0980062497"},{url:"/formList.svg",revision:"aef95c6f9019999cb9992dd2a21b3975"},{url:"/formSave.svg",revision:"e504c19b5bbbe70b0d2c83c9a73927b8"},{url:"/formSent.svg",revision:"eca064a2079c76fd61ca20528716ceee"},{url:"/formUndo.svg",revision:"91f221fce7b16f414253f9f1873902b2"},{url:"/hauling.svg",revision:"57a718f3ba6755d80106a01a21063d79"},{url:"/haulingFilled.svg",revision:"ebe07b7f9ca47cc45f7fd6b48e6d02ca"},{url:"/header.svg",revision:"479adabc8b70a52625854fba7e433548"},{url:"/home-white.svg",revision:"50f6d70fd4e0a491044340d69138b394"},{url:"/home.svg",revision:"c78e854f7b56bb952a0a22fc1980bc7d"},{url:"/icon-192X192.png",revision:"47aa596be976071dfff9c1e5455ecc80"},{url:"/icon-384x384.png",revision:"99d1987d20b874836d950b95586968ae"},{url:"/icon-512x512.png",revision:"c16feee422d814c946012979c0d02f00"},{url:"/inbox-white.svg",revision:"dfb451099519b395ed3d9b15daad5708"},{url:"/inbox.svg",revision:"03332320e762076df175dde846d62022"},{url:"/inboxFilled.svg",revision:"8b5e6d06de80a9927ae85007277b98a9"},{url:"/information.svg",revision:"3e9a1c81c8fd396fdf6849975d62dbb5"},{url:"/injury.svg",revision:"18121c3e1dc867f5d0a332a7133057f4"},{url:"/ios/100.png",revision:"5222167951b54f250720976e5798bd81"},{url:"/ios/1024.png",revision:"4046bc171d373187ec11f28c06ceb334"},{url:"/ios/114.png",revision:"2aec0491cbf7eb22d34f3079cf2eff54"},{url:"/ios/120.png",revision:"52ac3d673f239d4c822b20c337c73db2"},{url:"/ios/128.png",revision:"bc9dd541b1a0aaff432df4f5b0fb88b4"},{url:"/ios/144.png",revision:"554cb47f9b447d7c049e0af337f30598"},{url:"/ios/152.png",revision:"3085835ae98a412b5f8bb41b3d7fccae"},{url:"/ios/16.png",revision:"876eb7a741ae46958961bf0ed6c30ba0"},{url:"/ios/167.png",revision:"a7230785a2891b4659792d7d6cab6ffb"},{url:"/ios/180.png",revision:"5cdb37c8967955e3a4fae232138f3c03"},{url:"/ios/192.png",revision:"c7acce3a1ca251dc273817cd5462a666"},{url:"/ios/20.png",revision:"8c0f62f1ddd74accb89e318fe432a70f"},{url:"/ios/256.png",revision:"19665fd60a705b586e5cc5c93eddc1f2"},{url:"/ios/29.png",revision:"536c6f0fcf359bba5e0b45c9bb4fc260"},{url:"/ios/32.png",revision:"55a60f12c21ea3fffa61d8f7a34cfc33"},{url:"/ios/40.png",revision:"a38e8ea2895f2cfe54be4ce7465e7952"},{url:"/ios/50.png",revision:"db849275edb671b9ea7c45ef6f4fc489"},{url:"/ios/512.png",revision:"daaca9c8424139dcfdeed3181b0a7cbd"},{url:"/ios/57.png",revision:"8511d6a6a25878417899d2041de166ff"},{url:"/ios/58.png",revision:"001948872d9d885d20140b7aa9cc6ff8"},{url:"/ios/60.png",revision:"80209b85fb0be10562bcee31908d2a95"},{url:"/ios/64.png",revision:"066fa2094719e23427225f5517889ed7"},{url:"/ios/72.png",revision:"0544816fbb562d0825feb6e9ff73edff"},{url:"/ios/76.png",revision:"0a1c5da7e5a9673345c44f7e1c87f0d4"},{url:"/ios/80.png",revision:"91b74cbd7d0f0119a164a0750462149e"},{url:"/ios/87.png",revision:"74c6a6624174372f1fc483515c820c2f"},{url:"/jobsite-white.svg",revision:"1d991bcf439a2538ed78c709c708954b"},{url:"/jobsite.svg",revision:"f3909b4ff16c4f25e211fb0dc7556050"},{url:"/key.svg",revision:"f39b7da7fb64da838435e228ca3113ed"},{url:"/language.svg",revision:"e4615572d416d5df80c463dc53d5e0b7"},{url:"/layout.svg",revision:"a3b50fb72e8fab8feac60ccbd864e31a"},{url:"/logo.svg",revision:"8e78aff787124cc6567c2ccfd520099d"},{url:"/logo_JPG.jpg",revision:"e84c38cbd142e140688beacd59e6a0cf"},{url:"/manifest.json",revision:"05d4e6a0d9c7c7bebec473d08ea582b2"},{url:"/mechanic.svg",revision:"4f39d9d1ff5c68a81b93c7f6611d8e9d"},{url:"/message.svg",revision:"c55f300ce914f4b28036e30580ee104c"},{url:"/mileage.svg",revision:"11e1eb1d01c9be54e3c8394e198db717"},{url:"/minus.svg",revision:"d2ec83afd3490cac9da747386df54184"},{url:"/moreOptions.svg",revision:"d3019cfec5d1102ec73147f1dd5e5faf"},{url:"/moreOptionsCircle.svg",revision:"02d86b2a2acd7707b4cef84da6ede884"},{url:"/number.svg",revision:"6f90bfd3f431fe8602537fe47b9a109b"},{url:"/plus-stroke-white.svg",revision:"f23aeed115ad4273e185d3e73a314ba7"},{url:"/plus-white.svg",revision:"a3d859abbf1ce862eb0aa1d070d85fbf"},{url:"/plus.svg",revision:"33d3fcf5251a72f1eef80aaa19f7ab55"},{url:"/policies.svg",revision:"25de43083a5e24cdf83a7ba62d6c28c4"},{url:"/priorityDelay.svg",revision:"5286ef619f78a44968b600b83bba55a0"},{url:"/priorityHigh.svg",revision:"fa44198aecf39d4748914d7f3b4a0d03"},{url:"/priorityLow.svg",revision:"11efcad98767f7c049d90ece2f425f1a"},{url:"/priorityMedium.svg",revision:"219c8a19b23f94db24162d544e1605ec"},{url:"/priorityPending.svg",revision:"45d3ef273daf2b4daadcb0b6d0064789"},{url:"/priorityToday.svg",revision:"91fec1dfbc8a1268488c12a9ef9fa4ac"},{url:"/profileEmpty.svg",revision:"66491e65dd6114a9acf6990181fc44d9"},{url:"/profileFilled.svg",revision:"ff2e851ef303666374501b541725d34c"},{url:"/qrCode-white.svg",revision:"0bcc19567973489fa8a15e759b729d5b"},{url:"/qrCode.svg",revision:"a97e34eddfd5bae153771e734fd96e4c"},{url:"/radio.svg",revision:"17ecc9b4e325da65fc947c71d4bc1ade"},{url:"/refuel.svg",revision:"07fcabd768272b84f61007c4e130b8fc"},{url:"/refuelFilled.svg",revision:"d75e2ee23202812499b36b050fff841c"},{url:"/searchLeft.svg",revision:"7a4c52a1de03e0fbe1f2f949fed97367"},{url:"/searchRight.svg",revision:"9ac605dce8646197fa3ae4c90d93f8c9"},{url:"/settingsFilled.svg",revision:"4b9876c68beb7435b569e5a5a0252501"},{url:"/shiftScanLogo.svg",revision:"f6cdbf4c760167995264751f68ebb2af"},{url:"/shiftscanlogoHorizontal.svg",revision:"f6dac5483d88c1cdebc5f92190cc1d2f"},{url:"/spinner.svg",revision:"7eedb8067da00e29baf517b3cbfffc43"},{url:"/star.svg",revision:"744181e34e34c3385a845fa37cf51454"},{url:"/starFilled.svg",revision:"5ad7cd14d1deca592924c230b5a39218"},{url:"/state.svg",revision:"83c029784c461361996615f83c3850ef"},{url:"/stateFilled.svg",revision:"183c533850d5a2540b71b2c44bef9d11"},{url:"/statusApproved.svg",revision:"0e767f6c9f2c8f20480d3120d6b6e198"},{url:"/statusApprovedFilled.svg",revision:"6fc76b4c67d20fa9ffbdb451a30e8071"},{url:"/statusDenied.svg",revision:"7b1167c6857f8169396ae88ffbce95e6"},{url:"/statusDeniedFilled.svg",revision:"52740140811df8fc4cd87604621c9960"},{url:"/statusOffline.svg",revision:"d1931f81ba9fe9342e78e181c2216aaa"},{url:"/statusOngoing-white.svg",revision:"d07cb3a9608fbfa3e64486e49eb74365"},{url:"/statusOngoing.svg",revision:"7eedb8067da00e29baf517b3cbfffc43"},{url:"/statusOngoingFilled.svg",revision:"5e18aea67f178cddbdc7e2e449ea6e3a"},{url:"/statusOnline.svg",revision:"ebed6ae0692535bcfee10ada59c779c3"},{url:"/statusUnfinished.svg",revision:"038f7353e12e76758c9e214925a07a0c"},{url:"/swe-worker-5c72df51bb1f6ee0.js",revision:"447aee4a2c24fefe7917563f0255d688"},{url:"/tasco.svg",revision:"1280d07dd40b7604d022f71c3728ccfb"},{url:"/team-white.svg",revision:"0142ac2317a2b44b141c270916ab32ca"},{url:"/team.svg",revision:"203a48a7519f73fe7f24fff1cae20204"},{url:"/timecards.svg",revision:"7771f6f39b0e3d775e88c6d542d2d68b"},{url:"/tinyCheckMark-white.svg",revision:"a6d6eadc3dbfaafbf1e8257346eee02c"},{url:"/title.svg",revision:"15b1291ddf60f0ceb36ce46377fa6dd3"},{url:"/trash-red.svg",revision:"a42dfceb5b388b58d0c41c13ee0c17f7"},{url:"/trash.svg",revision:"229905df8ec07dd400f7a23deb582f0c"},{url:"/trucking.svg",revision:"ab3794a1a3073bc2926a1978a038300c"},{url:"/user-white.svg",revision:"6f6e9f9b948a6dfde98ca0b41985bbc0"},{url:"/user.svg",revision:"8756e7b0a6bc7ba8068b0f0806b5e5e7"},{url:"/windows11/LargeTile.scale-100.png",revision:"e3ae3637973ea93ff81e73ee37effe0c"},{url:"/windows11/LargeTile.scale-125.png",revision:"6e4ec698528b71f25efd6ff2dd68e89f"},{url:"/windows11/LargeTile.scale-150.png",revision:"b636fa1812d337a4ca1224b433a3d855"},{url:"/windows11/LargeTile.scale-200.png",revision:"d18a045105d03ba23b624c109708f7bb"},{url:"/windows11/LargeTile.scale-400.png",revision:"d4176f9ebaacc0bbfebf56fd56e253ce"},{url:"/windows11/SmallTile.scale-100.png",revision:"bce1445e4fb91c8c0842780a4026522e"},{url:"/windows11/SmallTile.scale-125.png",revision:"9415b7ff287961c074dee1834b42aad9"},{url:"/windows11/SmallTile.scale-150.png",revision:"87d19729c39b8d355a62d180d91bb31b"},{url:"/windows11/SmallTile.scale-200.png",revision:"7c5db43e6077ca3c958c21c85bec6d25"},{url:"/windows11/SmallTile.scale-400.png",revision:"1913d6ed56bc5b31dc60858acbcace6f"},{url:"/windows11/SplashScreen.scale-100.png",revision:"a8215f24409e6231d1dc00a0907ed4a8"},{url:"/windows11/SplashScreen.scale-125.png",revision:"097ac942f0459d3481fbd62c0f97e299"},{url:"/windows11/SplashScreen.scale-150.png",revision:"201fc53647b740af27d9b004cafeaaf8"},{url:"/windows11/SplashScreen.scale-200.png",revision:"1b1081f94203f3823ed2dd1f8dabb29a"},{url:"/windows11/SplashScreen.scale-400.png",revision:"0d6ee6f415a7d2101792e2b8312af059"},{url:"/windows11/Square150x150Logo.scale-100.png",revision:"7b6855c0928dad5db20403341b059afc"},{url:"/windows11/Square150x150Logo.scale-125.png",revision:"dd8460f97de8d2cf5826468f8f8153db"},{url:"/windows11/Square150x150Logo.scale-150.png",revision:"8bfd4727bbeb856bd1af17e0cdef44ea"},{url:"/windows11/Square150x150Logo.scale-200.png",revision:"c95f1e579117442bcf6a388818260303"},{url:"/windows11/Square150x150Logo.scale-400.png",revision:"a20add32148ad6bfcc0d435ac4986eb5"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-16.png",revision:"90a9d6c44b09d44f416b62df26224428"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-20.png",revision:"779afd7d8c738f0a0c73a4a6d63e4aff"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-24.png",revision:"a65f2b674acb5c2af36db18e9b4a63f1"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-256.png",revision:"723b833ab63af25543dfe92f0a56e6d5"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-30.png",revision:"3a63be52b054d426a0ee4ce69b783e12"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-32.png",revision:"261d9c7ca92fefc2230721bb99cbcae1"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-36.png",revision:"a709411193e6374836f5fa82b8afa6c1"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-40.png",revision:"4251a65706de0f7c86b1bee205eb53a7"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-44.png",revision:"9f4cf9e0ba353f1bbb19fc9f6e40523a"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-48.png",revision:"c87a2b33f89afc0199219ea0b6e191ac"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-60.png",revision:"727adb65708020078e70655ef5444a7b"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-64.png",revision:"04915cfcb9f9ad29a6223fb193c79a14"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-72.png",revision:"1f5e7f8fb46e7d8e425ff06c34338b6d"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-80.png",revision:"3a7aa9e5f873375880d85aa0c2f51306"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-96.png",revision:"a151f9aefbcedac0d11aba5cf1108af8"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-16.png",revision:"90a9d6c44b09d44f416b62df26224428"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-20.png",revision:"779afd7d8c738f0a0c73a4a6d63e4aff"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-24.png",revision:"a65f2b674acb5c2af36db18e9b4a63f1"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-256.png",revision:"723b833ab63af25543dfe92f0a56e6d5"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-30.png",revision:"3a63be52b054d426a0ee4ce69b783e12"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-32.png",revision:"261d9c7ca92fefc2230721bb99cbcae1"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-36.png",revision:"a709411193e6374836f5fa82b8afa6c1"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-40.png",revision:"4251a65706de0f7c86b1bee205eb53a7"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-44.png",revision:"9f4cf9e0ba353f1bbb19fc9f6e40523a"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-48.png",revision:"c87a2b33f89afc0199219ea0b6e191ac"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-60.png",revision:"727adb65708020078e70655ef5444a7b"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-64.png",revision:"04915cfcb9f9ad29a6223fb193c79a14"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-72.png",revision:"1f5e7f8fb46e7d8e425ff06c34338b6d"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-80.png",revision:"3a7aa9e5f873375880d85aa0c2f51306"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-96.png",revision:"a151f9aefbcedac0d11aba5cf1108af8"},{url:"/windows11/Square44x44Logo.scale-100.png",revision:"9f4cf9e0ba353f1bbb19fc9f6e40523a"},{url:"/windows11/Square44x44Logo.scale-125.png",revision:"95498060191b331348eb8f5b5ab6dfe7"},{url:"/windows11/Square44x44Logo.scale-150.png",revision:"aced07a7a4627cdad487ed0a46faeebb"},{url:"/windows11/Square44x44Logo.scale-200.png",revision:"db11fa1c0b22e2344ecf9dea9af68a2c"},{url:"/windows11/Square44x44Logo.scale-400.png",revision:"5f1b23fb222c880bf3c533643f43b4ca"},{url:"/windows11/Square44x44Logo.targetsize-16.png",revision:"90a9d6c44b09d44f416b62df26224428"},{url:"/windows11/Square44x44Logo.targetsize-20.png",revision:"779afd7d8c738f0a0c73a4a6d63e4aff"},{url:"/windows11/Square44x44Logo.targetsize-24.png",revision:"a65f2b674acb5c2af36db18e9b4a63f1"},{url:"/windows11/Square44x44Logo.targetsize-256.png",revision:"723b833ab63af25543dfe92f0a56e6d5"},{url:"/windows11/Square44x44Logo.targetsize-30.png",revision:"3a63be52b054d426a0ee4ce69b783e12"},{url:"/windows11/Square44x44Logo.targetsize-32.png",revision:"261d9c7ca92fefc2230721bb99cbcae1"},{url:"/windows11/Square44x44Logo.targetsize-36.png",revision:"a709411193e6374836f5fa82b8afa6c1"},{url:"/windows11/Square44x44Logo.targetsize-40.png",revision:"4251a65706de0f7c86b1bee205eb53a7"},{url:"/windows11/Square44x44Logo.targetsize-44.png",revision:"9f4cf9e0ba353f1bbb19fc9f6e40523a"},{url:"/windows11/Square44x44Logo.targetsize-48.png",revision:"c87a2b33f89afc0199219ea0b6e191ac"},{url:"/windows11/Square44x44Logo.targetsize-60.png",revision:"727adb65708020078e70655ef5444a7b"},{url:"/windows11/Square44x44Logo.targetsize-64.png",revision:"04915cfcb9f9ad29a6223fb193c79a14"},{url:"/windows11/Square44x44Logo.targetsize-72.png",revision:"1f5e7f8fb46e7d8e425ff06c34338b6d"},{url:"/windows11/Square44x44Logo.targetsize-80.png",revision:"3a7aa9e5f873375880d85aa0c2f51306"},{url:"/windows11/Square44x44Logo.targetsize-96.png",revision:"a151f9aefbcedac0d11aba5cf1108af8"},{url:"/windows11/StoreLogo.scale-100.png",revision:"db849275edb671b9ea7c45ef6f4fc489"},{url:"/windows11/StoreLogo.scale-125.png",revision:"728e6527e2d77371e3007caf007b116f"},{url:"/windows11/StoreLogo.scale-150.png",revision:"9224f9a45257c2103e10aa9026af46d1"},{url:"/windows11/StoreLogo.scale-200.png",revision:"5222167951b54f250720976e5798bd81"},{url:"/windows11/StoreLogo.scale-400.png",revision:"6562c92cea67a6232c5727e464a9dd83"},{url:"/windows11/Wide310x150Logo.scale-100.png",revision:"7fc14247c4af58216f16d5f4f98fc46b"},{url:"/windows11/Wide310x150Logo.scale-125.png",revision:"a0c8d4ccacb2bc65438458df95ab3d60"},{url:"/windows11/Wide310x150Logo.scale-150.png",revision:"34161357c0cdafe2651b8d197c019798"},{url:"/windows11/Wide310x150Logo.scale-200.png",revision:"a8215f24409e6231d1dc00a0907ed4a8"},{url:"/windows11/Wide310x150Logo.scale-400.png",revision:"1b1081f94203f3823ed2dd1f8dabb29a"}],{ignoreURLParametersMatching:[/^utm_/,/^fbclid$/]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({response:e})=>e&&"opaqueredirect"===e.type?new Response(e.body,{status:200,statusText:"OK",headers:e.headers}):e}]}),"GET"),e.registerRoute(/^https?.*/,new e.NetworkFirst({cacheName:"offlineCache",plugins:[new e.ExpirationPlugin({maxEntries:200,maxAgeSeconds:86400})]}),"GET"),self.__WB_DISABLE_DEV_LOGS=!0}));
+if(!self.define){let e,a={};const i=(i,s)=>(i=new URL(i+".js",s).href,a[i]||new Promise((a=>{if("document"in self){const e=document.createElement("script");e.src=i,e.onload=a,document.head.appendChild(e)}else e=i,importScripts(i),a()})).then((()=>{let e=a[i];if(!e)throw new Error(`Module ${i} didn’t register its module`);return e})));self.define=(s,c)=>{const r=e||("document"in self?document.currentScript.src:"")||location.href;if(a[r])return;let n={};const t=e=>i(e,r),d={module:{uri:r},exports:n,require:t};a[r]=Promise.all(s.map((e=>d[e]||t(e)))).then((e=>(c(...e),n)))}}define(["./workbox-31da5539"],(function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/Settings.svg",revision:"ebe8b6a8bff9bd2fe8669ad036e8edf8"},{url:"/_next/static/GuMQR4V94FSL4iRL6Osef/_buildManifest.js",revision:"bc2675d22d82f5f5bd07c2ba930f02b2"},{url:"/_next/static/GuMQR4V94FSL4iRL6Osef/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/chunks/1095-f78c68c0229f4b99.js",revision:"f78c68c0229f4b99"},{url:"/_next/static/chunks/1179-e86479f27467d923.js",revision:"e86479f27467d923"},{url:"/_next/static/chunks/1317-5bb21e343418ec0f.js",revision:"5bb21e343418ec0f"},{url:"/_next/static/chunks/1401-200e2839e2bf3cec.js",revision:"200e2839e2bf3cec"},{url:"/_next/static/chunks/1692-cc16fa14e0e5bbf3.js",revision:"cc16fa14e0e5bbf3"},{url:"/_next/static/chunks/2019-244f4105c4691c0c.js",revision:"244f4105c4691c0c"},{url:"/_next/static/chunks/2057-668deaa0eb77c6a4.js",revision:"668deaa0eb77c6a4"},{url:"/_next/static/chunks/2093-162efb4b28b6a687.js",revision:"162efb4b28b6a687"},{url:"/_next/static/chunks/2170a4aa-8c96009f60e20411.js",revision:"8c96009f60e20411"},{url:"/_next/static/chunks/2328-d3a6681e5c4f0896.js",revision:"d3a6681e5c4f0896"},{url:"/_next/static/chunks/2403.ca107014ddba66de.js",revision:"ca107014ddba66de"},{url:"/_next/static/chunks/2443-261193c0aabf52ad.js",revision:"261193c0aabf52ad"},{url:"/_next/static/chunks/2541-b9deac5a6c87ddb9.js",revision:"b9deac5a6c87ddb9"},{url:"/_next/static/chunks/2551-2b96b83195377356.js",revision:"2b96b83195377356"},{url:"/_next/static/chunks/2722-48cde404f78497eb.js",revision:"48cde404f78497eb"},{url:"/_next/static/chunks/3068-e80cb9e7cb7a36bb.js",revision:"e80cb9e7cb7a36bb"},{url:"/_next/static/chunks/3206-67be7e0a1bd37815.js",revision:"67be7e0a1bd37815"},{url:"/_next/static/chunks/3328-1e73249e263f1cb2.js",revision:"1e73249e263f1cb2"},{url:"/_next/static/chunks/3713-01fa541b1020dd9d.js",revision:"01fa541b1020dd9d"},{url:"/_next/static/chunks/3782-7f7d9d67197b4aca.js",revision:"7f7d9d67197b4aca"},{url:"/_next/static/chunks/4095-8d931f3f0218645a.js",revision:"8d931f3f0218645a"},{url:"/_next/static/chunks/4114-b838e9dd1d2a7f67.js",revision:"b838e9dd1d2a7f67"},{url:"/_next/static/chunks/4493-4e6be3f951868868.js",revision:"4e6be3f951868868"},{url:"/_next/static/chunks/4914-8e3585d0f1c9f02d.js",revision:"8e3585d0f1c9f02d"},{url:"/_next/static/chunks/4bd1b696-88e37c56edce0f59.js",revision:"88e37c56edce0f59"},{url:"/_next/static/chunks/5152-9e6b4080454ffa3c.js",revision:"9e6b4080454ffa3c"},{url:"/_next/static/chunks/52774a7f-9f526229934f791c.js",revision:"9f526229934f791c"},{url:"/_next/static/chunks/5395-71cc1bac006bc080.js",revision:"71cc1bac006bc080"},{url:"/_next/static/chunks/5493-002a49a6d796cca8.js",revision:"002a49a6d796cca8"},{url:"/_next/static/chunks/5535-ed37b1f77860aa26.js",revision:"ed37b1f77860aa26"},{url:"/_next/static/chunks/5705-2d0aee81a02a655e.js",revision:"2d0aee81a02a655e"},{url:"/_next/static/chunks/5751-277f95667565cf9f.js",revision:"277f95667565cf9f"},{url:"/_next/static/chunks/6323-d1e244e417681874.js",revision:"d1e244e417681874"},{url:"/_next/static/chunks/6544-386355f8b296dc20.js",revision:"386355f8b296dc20"},{url:"/_next/static/chunks/6612-126530b7baba5f69.js",revision:"126530b7baba5f69"},{url:"/_next/static/chunks/6671-c6ce7faab2868fe3.js",revision:"c6ce7faab2868fe3"},{url:"/_next/static/chunks/6697-68520a497e02947b.js",revision:"68520a497e02947b"},{url:"/_next/static/chunks/6862-81cfc3fff123ab8a.js",revision:"81cfc3fff123ab8a"},{url:"/_next/static/chunks/6874-bf3fa41ffddf2fd7.js",revision:"bf3fa41ffddf2fd7"},{url:"/_next/static/chunks/7111-088fe5dbd51639ef.js",revision:"088fe5dbd51639ef"},{url:"/_next/static/chunks/7253-b6a7c4b8ba2a4ccc.js",revision:"b6a7c4b8ba2a4ccc"},{url:"/_next/static/chunks/7655-e651b79933d954e7.js",revision:"e651b79933d954e7"},{url:"/_next/static/chunks/7686-ba6a1435558ab5a4.js",revision:"ba6a1435558ab5a4"},{url:"/_next/static/chunks/789-86a3b92415cd819c.js",revision:"86a3b92415cd819c"},{url:"/_next/static/chunks/822-0edc64ee089dcd0d.js",revision:"0edc64ee089dcd0d"},{url:"/_next/static/chunks/8310-1a53d637c8ce85b0.js",revision:"1a53d637c8ce85b0"},{url:"/_next/static/chunks/8370-4fedd68255eb66d3.js",revision:"4fedd68255eb66d3"},{url:"/_next/static/chunks/8436.fe8b1513f8e56bfa.js",revision:"fe8b1513f8e56bfa"},{url:"/_next/static/chunks/8743.e58f7147681da72a.js",revision:"e58f7147681da72a"},{url:"/_next/static/chunks/9127-f0ee2a66d843196b.js",revision:"f0ee2a66d843196b"},{url:"/_next/static/chunks/9313-eae8d054304ecf16.js",revision:"eae8d054304ecf16"},{url:"/_next/static/chunks/9373-12ddcc70f6722621.js",revision:"12ddcc70f6722621"},{url:"/_next/static/chunks/9481-4cf6b57b1f1291d3.js",revision:"4cf6b57b1f1291d3"},{url:"/_next/static/chunks/9648-8dba046886ed5b99.js",revision:"8dba046886ed5b99"},{url:"/_next/static/chunks/9763-3b7d58f5c287c919.js",revision:"3b7d58f5c287c919"},{url:"/_next/static/chunks/9872-7c697f701ea3a2f2.js",revision:"7c697f701ea3a2f2"},{url:"/_next/static/chunks/app/(routes)/admins/clients/page-47cc4fbfa5d49bab.js",revision:"47cc4fbfa5d49bab"},{url:"/_next/static/chunks/app/(routes)/admins/cost-codes/page-2a6baf7353718af8.js",revision:"2a6baf7353718af8"},{url:"/_next/static/chunks/app/(routes)/admins/crew/page-05ce66c8ddfd6f84.js",revision:"05ce66c8ddfd6f84"},{url:"/_next/static/chunks/app/(routes)/admins/equipment/page-888f17bcfdbf9a80.js",revision:"888f17bcfdbf9a80"},{url:"/_next/static/chunks/app/(routes)/admins/forms/%5Bid%5D/page-949c5508a7ead88b.js",revision:"949c5508a7ead88b"},{url:"/_next/static/chunks/app/(routes)/admins/forms/create/page-5efe530cd05cb6db.js",revision:"5efe530cd05cb6db"},{url:"/_next/static/chunks/app/(routes)/admins/forms/edit/%5Bid%5D/page-c6b5488e0a16dce1.js",revision:"c6b5488e0a16dce1"},{url:"/_next/static/chunks/app/(routes)/admins/forms/page-d38267fe3524461a.js",revision:"d38267fe3524461a"},{url:"/_next/static/chunks/app/(routes)/admins/jobsites/page-6b6077e6021333a9.js",revision:"6b6077e6021333a9"},{url:"/_next/static/chunks/app/(routes)/admins/layout-74b03a55699d8958.js",revision:"74b03a55699d8958"},{url:"/_next/static/chunks/app/(routes)/admins/page-cdc3b85b0c3243d3.js",revision:"cdc3b85b0c3243d3"},{url:"/_next/static/chunks/app/(routes)/admins/personnel/page-21f0fa8f2d605eb4.js",revision:"21f0fa8f2d605eb4"},{url:"/_next/static/chunks/app/(routes)/admins/reports/page-4c56b8e8379642e8.js",revision:"4c56b8e8379642e8"},{url:"/_next/static/chunks/app/(routes)/admins/timesheets/page-2fe4dd552031c580.js",revision:"2fe4dd552031c580"},{url:"/_next/static/chunks/app/(routes)/break/page-21caaf8314224166.js",revision:"21caaf8314224166"},{url:"/_next/static/chunks/app/(routes)/clock/page-916e93cebf243368.js",revision:"916e93cebf243368"},{url:"/_next/static/chunks/app/(routes)/dashboard/clock-out/(components)/log/page-b6e632fe83e2498b.js",revision:"b6e632fe83e2498b"},{url:"/_next/static/chunks/app/(routes)/dashboard/clock-out/page-df61dcb16cfc00f3.js",revision:"df61dcb16cfc00f3"},{url:"/_next/static/chunks/app/(routes)/dashboard/equipment/%5Bid%5D/page-24ccbf9235edbe67.js",revision:"24ccbf9235edbe67"},{url:"/_next/static/chunks/app/(routes)/dashboard/equipment/log-new/page-0e751569e3f67681.js",revision:"0e751569e3f67681"},{url:"/_next/static/chunks/app/(routes)/dashboard/equipment/page-01c72909cb40becf.js",revision:"01c72909cb40becf"},{url:"/_next/static/chunks/app/(routes)/dashboard/mechanic/edit-repair-details/%5Bid%5D/page-56d0aeae424b938f.js",revision:"56d0aeae424b938f"},{url:"/_next/static/chunks/app/(routes)/dashboard/mechanic/new-repair/page-490823836af9207c.js",revision:"490823836af9207c"},{url:"/_next/static/chunks/app/(routes)/dashboard/mechanic/page-341d41682cc8b6a6.js",revision:"341d41682cc8b6a6"},{url:"/_next/static/chunks/app/(routes)/dashboard/mechanic/projects/%5Bid%5D/page-55aeee7523b72689.js",revision:"55aeee7523b72689"},{url:"/_next/static/chunks/app/(routes)/dashboard/myTeam/%5Bid%5D/employee/%5BemployeeId%5D/page-0dda2ae04de3ded2.js",revision:"0dda2ae04de3ded2"},{url:"/_next/static/chunks/app/(routes)/dashboard/myTeam/%5Bid%5D/page-ccb5d2e7958313d7.js",revision:"ccb5d2e7958313d7"},{url:"/_next/static/chunks/app/(routes)/dashboard/myTeam/page-73437efe4ff35b75.js",revision:"73437efe4ff35b75"},{url:"/_next/static/chunks/app/(routes)/dashboard/myTeam/timecards/page-9eb9b56fc9ccb487.js",revision:"9eb9b56fc9ccb487"},{url:"/_next/static/chunks/app/(routes)/dashboard/page-4016bf3e76d83acb.js",revision:"4016bf3e76d83acb"},{url:"/_next/static/chunks/app/(routes)/dashboard/qr-generator/add-equipment/page-92380df793993c51.js",revision:"92380df793993c51"},{url:"/_next/static/chunks/app/(routes)/dashboard/qr-generator/add-jobsite/page-bf715c823804df6f.js",revision:"bf715c823804df6f"},{url:"/_next/static/chunks/app/(routes)/dashboard/qr-generator/page-76d6385b787361da.js",revision:"76d6385b787361da"},{url:"/_next/static/chunks/app/(routes)/dashboard/switch-jobs/page-c6da0ba6d1e256f0.js",revision:"c6da0ba6d1e256f0"},{url:"/_next/static/chunks/app/(routes)/dashboard/tasco/page-8a3f0c2b107e2d45.js",revision:"8a3f0c2b107e2d45"},{url:"/_next/static/chunks/app/(routes)/dashboard/truckingAssistant/page-2ccb8d75e7fb553e.js",revision:"2ccb8d75e7fb553e"},{url:"/_next/static/chunks/app/(routes)/hamburger/changePassword/page-38d824dd29b7f69c.js",revision:"38d824dd29b7f69c"},{url:"/_next/static/chunks/app/(routes)/hamburger/inbox/formSubmission/%5Bid%5D/page-50a86d5272819b5b.js",revision:"50a86d5272819b5b"},{url:"/_next/static/chunks/app/(routes)/hamburger/inbox/page-123a92a845220059.js",revision:"123a92a845220059"},{url:"/_next/static/chunks/app/(routes)/hamburger/profile/page-70cbae58e9e16cdb.js",revision:"70cbae58e9e16cdb"},{url:"/_next/static/chunks/app/(routes)/information/page-89d7583d61422b6d.js",revision:"89d7583d61422b6d"},{url:"/_next/static/chunks/app/(routes)/signin/forgot-password/page-452ac6e4dbf00e2c.js",revision:"452ac6e4dbf00e2c"},{url:"/_next/static/chunks/app/(routes)/signin/new-password/page-48db32da7718445f.js",revision:"48db32da7718445f"},{url:"/_next/static/chunks/app/(routes)/signin/page-776eda5d900bfbb8.js",revision:"776eda5d900bfbb8"},{url:"/_next/static/chunks/app/(routes)/signin/signup/page-fd3ccee82d025100.js",revision:"fd3ccee82d025100"},{url:"/_next/static/chunks/app/(routes)/timesheets/page-770ac7911f8bc820.js",revision:"770ac7911f8bc820"},{url:"/_next/static/chunks/app/_not-found/page-e32296c53d3e4077.js",revision:"e32296c53d3e4077"},{url:"/_next/static/chunks/app/api/(admins)/crewManager/route-a6ad4fd50bc07e88.js",revision:"a6ad4fd50bc07e88"},{url:"/_next/static/chunks/app/api/(admins)/employeeInfo/%5Bemployee%5D/route-b8fc451918d6fb9e.js",revision:"b8fc451918d6fb9e"},{url:"/_next/static/chunks/app/api/(admins)/getAllActiveEmployeeName/route-ad31be4f725c1a32.js",revision:"ad31be4f725c1a32"},{url:"/_next/static/chunks/app/api/(admins)/getAllCostCodesByJobSites/route-0013863ff3a6e8dc.js",revision:"0013863ff3a6e8dc"},{url:"/_next/static/chunks/app/api/(admins)/getAllCrews/route-8a68e32a18b4e2f1.js",revision:"8a68e32a18b4e2f1"},{url:"/_next/static/chunks/app/api/(admins)/getAllEmployees/route-96c691df49827c48.js",revision:"96c691df49827c48"},{url:"/_next/static/chunks/app/api/(admins)/getAllEquipment/route-47258141a670c4a4.js",revision:"47258141a670c4a4"},{url:"/_next/static/chunks/app/api/(admins)/getAllForms/route-5ed55ec535bd3c90.js",revision:"5ed55ec535bd3c90"},{url:"/_next/static/chunks/app/api/(admins)/getAllTimesheetInfo/route-3b13943fe248499b.js",revision:"3b13943fe248499b"},{url:"/_next/static/chunks/app/api/(admins)/getAllTimesheetsPending/route-bdae5e866282ceef.js",revision:"bdae5e866282ceef"},{url:"/_next/static/chunks/app/api/(admins)/getClientById/%5Bid%5D/route-c2cccbabe90740c9.js",revision:"c2cccbabe90740c9"},{url:"/_next/static/chunks/app/api/(admins)/getClientsDetails/route-aaca05f497bbcc55.js",revision:"aaca05f497bbcc55"},{url:"/_next/static/chunks/app/api/(admins)/getCostCodeDetails/route-813fb26c1b04b60f.js",revision:"813fb26c1b04b60f"},{url:"/_next/static/chunks/app/api/(admins)/getCostCodeSummary/route-90bd8f9764c1e92e.js",revision:"90bd8f9764c1e92e"},{url:"/_next/static/chunks/app/api/(admins)/getCostCodeTags/%5Bid%5D/route-f1aa8ee55e03ce8e.js",revision:"f1aa8ee55e03ce8e"},{url:"/_next/static/chunks/app/api/(admins)/getCrewByIdAdmin/%5Bid%5D/route-6e363c5f3aef3c06.js",revision:"6e363c5f3aef3c06"},{url:"/_next/static/chunks/app/api/(admins)/getDashboard/route-be74688f5dce0471.js",revision:"be74688f5dce0471"},{url:"/_next/static/chunks/app/api/(admins)/getEquipmentByEquipmentId/%5Bequipment%5D/route-f843dc2e35581a19.js",revision:"f843dc2e35581a19"},{url:"/_next/static/chunks/app/api/(admins)/getEquipmentDetails/route-aed522d0f2ca28f9.js",revision:"aed522d0f2ca28f9"},{url:"/_next/static/chunks/app/api/(admins)/getEquipmentSummary/route-e9b56c06596eedaf.js",revision:"e9b56c06596eedaf"},{url:"/_next/static/chunks/app/api/(admins)/getFormSubmissionsById/%5Bid%5D/route-0b3a0cd9de35ee70.js",revision:"0b3a0cd9de35ee70"},{url:"/_next/static/chunks/app/api/(admins)/getForms/%5Bid%5D/route-82ece0b3426108d2.js",revision:"82ece0b3426108d2"},{url:"/_next/static/chunks/app/api/(admins)/getJobsiteById/%5Bid%5D/route-a4ca1a73b8757ec1.js",revision:"a4ca1a73b8757ec1"},{url:"/_next/static/chunks/app/api/(admins)/getJobsiteSummary/route-65701a5a0abb06ad.js",revision:"65701a5a0abb06ad"},{url:"/_next/static/chunks/app/api/(admins)/getMechanicProjectSummary/route-c843295f63c4e44f.js",revision:"c843295f63c4e44f"},{url:"/_next/static/chunks/app/api/(admins)/getTagById/%5Bid%5D/route-3a8d0f9ef7cb2ced.js",revision:"3a8d0f9ef7cb2ced"},{url:"/_next/static/chunks/app/api/(admins)/getTagSummary/route-43391b409c40c428.js",revision:"43391b409c40c428"},{url:"/_next/static/chunks/app/api/(admins)/getTimesheetById/route-e8b48214da8d3a54.js",revision:"e8b48214da8d3a54"},{url:"/_next/static/chunks/app/api/(admins)/jobsiteManager/route-9a4b336f15428190.js",revision:"9a4b336f15428190"},{url:"/_next/static/chunks/app/api/(admins)/personnelManager/route-02f9e90102c4e564.js",revision:"02f9e90102c4e564"},{url:"/_next/static/chunks/app/api/(admins)/reports/tasco/route-46a2f5328cf22bf1.js",revision:"46a2f5328cf22bf1"},{url:"/_next/static/chunks/app/api/(admins)/reports/truckingReport/route-4a88ee9b71036746.js",revision:"4a88ee9b71036746"},{url:"/_next/static/chunks/app/api/(clock)/getEquipmentList/route-4730f2eef80c4bcc.js",revision:"4730f2eef80c4bcc"},{url:"/_next/static/chunks/app/api/(clock)/getLogs/route-41e2b14cc47b8324.js",revision:"41e2b14cc47b8324"},{url:"/_next/static/chunks/app/api/(clock)/getMyTeamsUsers/route-1de587d2af723a09.js",revision:"1de587d2af723a09"},{url:"/_next/static/chunks/app/api/(clock)/getPendingTeamTimesheets/%5BcrewMembers%5D/route-030c108b1524f10d.js",revision:"030c108b1524f10d"},{url:"/_next/static/chunks/app/api/(clock)/getRecentCostCodes/route-66011e59ba798626.js",revision:"66011e59ba798626"},{url:"/_next/static/chunks/app/api/(clock)/getTodaysTimesheets/route-a7014ca9e8477ab2.js",revision:"a7014ca9e8477ab2"},{url:"/_next/static/chunks/app/api/(clock)/getTruckData/route-d43e55d8abea453d.js",revision:"d43e55d8abea453d"},{url:"/_next/static/chunks/app/api/(equipment)/getAllEquipmentIdAndQrId/route-c8f9906ad1ec689e.js",revision:"c8f9906ad1ec689e"},{url:"/_next/static/chunks/app/api/(equipment)/getCheckedList/route-6aa444051fc09b44.js",revision:"6aa444051fc09b44"},{url:"/_next/static/chunks/app/api/(equipment)/getEqUserLogs/%5Bid%5D/route-51627412ebb068bc.js",revision:"51627412ebb068bc"},{url:"/_next/static/chunks/app/api/(equipment)/getEqUserLogs/route-e9010271cbf63852.js",revision:"e9010271cbf63852"},{url:"/_next/static/chunks/app/api/(equipment)/getEquipmentRefueledLogs/%5Bid%5D/route-5c254f91ac2260c4.js",revision:"5c254f91ac2260c4"},{url:"/_next/static/chunks/app/api/(equipment)/getRecentJobDetails/route-b51c51dccc4f5d98.js",revision:"b51c51dccc4f5d98"},{url:"/_next/static/chunks/app/api/(forms)/employeeRequests/%5Bfilter%5D/route-108c4aea7f16e672.js",revision:"108c4aea7f16e672"},{url:"/_next/static/chunks/app/api/(forms)/form/%5Bid%5D/route-def5d21120e4c1c6.js",revision:"def5d21120e4c1c6"},{url:"/_next/static/chunks/app/api/(forms)/formDraft/%5Bid%5D/route-c27f4097ff3cec17.js",revision:"c27f4097ff3cec17"},{url:"/_next/static/chunks/app/api/(forms)/formDrafts/route-0c1896c75aa610df.js",revision:"0c1896c75aa610df"},{url:"/_next/static/chunks/app/api/(forms)/formSubmission/%5Bid%5D/route-4a77a032c874ef51.js",revision:"4a77a032c874ef51"},{url:"/_next/static/chunks/app/api/(forms)/formSubmissions/%5Bstatus%5D/route-2b172217bfd31857.js",revision:"2b172217bfd31857"},{url:"/_next/static/chunks/app/api/(forms)/forms/route-386e9b6eadce202b.js",revision:"386e9b6eadce202b"},{url:"/_next/static/chunks/app/api/(forms)/getDocumentById/%5Bid%5D/route-16a72506935bb68b.js",revision:"16a72506935bb68b"},{url:"/_next/static/chunks/app/api/(forms)/getDocumentTags/route-0d1d2efd90b7d4a0.js",revision:"0d1d2efd90b7d4a0"},{url:"/_next/static/chunks/app/api/(forms)/getDocuments/route-5735936db823474e.js",revision:"5735936db823474e"},{url:"/_next/static/chunks/app/api/(forms)/getEmployees/route-85b42aa19cf4b68e.js",revision:"85b42aa19cf4b68e"},{url:"/_next/static/chunks/app/api/(forms)/getEqForDocs/route-71e475a4876e7ff9.js",revision:"71e475a4876e7ff9"},{url:"/_next/static/chunks/app/api/(forms)/managerFormApproval/%5Bid%5D/route-4ad2749ba3cb8a8e.js",revision:"4ad2749ba3cb8a8e"},{url:"/_next/static/chunks/app/api/(forms)/teamSubmission/%5Bid%5D/route-e2fcfe03875ad2de.js",revision:"e2fcfe03875ad2de"},{url:"/_next/static/chunks/app/api/(general)/cookies/route-f198ae2c75bb0add.js",revision:"f198ae2c75bb0add"},{url:"/_next/static/chunks/app/api/(general)/getAllJobsites/route-4f2e0626ab8c78d2.js",revision:"4f2e0626ab8c78d2"},{url:"/_next/static/chunks/app/api/(general)/getAllTags/route-a8ba9526a1d939dd.js",revision:"a8ba9526a1d939dd"},{url:"/_next/static/chunks/app/api/(general)/getBannerData/route-ce96a4f79060d576.js",revision:"ce96a4f79060d576"},{url:"/_next/static/chunks/app/api/(general)/getCostCodes/route-839dc537fcbbd0ae.js",revision:"839dc537fcbbd0ae"},{url:"/_next/static/chunks/app/api/(general)/getCrewByCrewId/%5Bcrew%5D/route-04816e0d35252339.js",revision:"04816e0d35252339"},{url:"/_next/static/chunks/app/api/(general)/getEquipment/route-6f063fe970202a3c.js",revision:"6f063fe970202a3c"},{url:"/_next/static/chunks/app/api/(general)/getPayPeriodTimeSheets/route-ee7337de72860ecc.js",revision:"ee7337de72860ecc"},{url:"/_next/static/chunks/app/api/(general)/getRecentEquipment/route-bda9e71b194a8724.js",revision:"bda9e71b194a8724"},{url:"/_next/static/chunks/app/api/(general)/getRecentJobsites/route-5fc7e4a03a80fef1.js",revision:"5fc7e4a03a80fef1"},{url:"/_next/static/chunks/app/api/(general)/getRecentTimecard/route-63c4a5bb3607cfba.js",revision:"63c4a5bb3607cfba"},{url:"/_next/static/chunks/app/api/(general)/getRecentTimecardReturn/route-c4a891e6b10397d8.js",revision:"c4a891e6b10397d8"},{url:"/_next/static/chunks/app/api/(general)/getSettings/route-25a1789704590a17.js",revision:"25a1789704590a17"},{url:"/_next/static/chunks/app/api/(general)/getTimesheets/%5Bemployee%5D/route-d19db3e142dc44e4.js",revision:"d19db3e142dc44e4"},{url:"/_next/static/chunks/app/api/(general)/getTimesheets/route-e038697331997d18.js",revision:"e038697331997d18"},{url:"/_next/static/chunks/app/api/(general)/getUserSignature/route-51aaa98dc7282071.js",revision:"51aaa98dc7282071"},{url:"/_next/static/chunks/app/api/(mechanic)/getMaintenanceLogs/%5Bid%5D/route-bd938afe60677879.js",revision:"bd938afe60677879"},{url:"/_next/static/chunks/app/api/(mechanic)/getMaintenanceProjects/route-18b286a2359e9dfe.js",revision:"18b286a2359e9dfe"},{url:"/_next/static/chunks/app/api/(mechanic)/getReceivedInfo/%5Bid%5D/route-7544b71bd88555ea.js",revision:"7544b71bd88555ea"},{url:"/_next/static/chunks/app/api/(mechanic)/getRepairDetails/%5Bid%5D/route-089ac4bd0da69a42.js",revision:"089ac4bd0da69a42"},{url:"/_next/static/chunks/app/api/(myTeam)/getCrew/route-9d1d447937f92205.js",revision:"9d1d447937f92205"},{url:"/_next/static/chunks/app/api/(myTeam)/getCrewById/%5BcrewId%5D/route-97af907f88f480fa.js",revision:"97af907f88f480fa"},{url:"/_next/static/chunks/app/api/(myTeam)/getPendingTeamTimeSheets/route-8aa5bcff4e5e0ec6.js",revision:"8aa5bcff4e5e0ec6"},{url:"/_next/static/chunks/app/api/(myTeam)/getTeam/route-1d9bbd33c8160a21.js",revision:"1d9bbd33c8160a21"},{url:"/_next/static/chunks/app/api/(myTeam)/getUserInfo/%5Bid%5D/route-d3d91f6dee197267.js",revision:"d3d91f6dee197267"},{url:"/_next/static/chunks/app/api/(qrGenerator)/getEquipmentbyQrId/%5BqrId%5D/route-ee6b4690eb4b1acb.js",revision:"ee6b4690eb4b1acb"},{url:"/_next/static/chunks/app/api/(qrGenerator)/getJobsites/route-ec72fb07766b4045.js",revision:"ec72fb07766b4045"},{url:"/_next/static/chunks/app/api/(reviewWithDevun)/getSignature/%5Bid%5D/route-690be204793a0258.js",revision:"690be204793a0258"},{url:"/_next/static/chunks/app/api/(reviewWithDevun)/getTimesheetById/%5Bemployee%5D/route-a1410e3a9fd6a75d.js",revision:"a1410e3a9fd6a75d"},{url:"/_next/static/chunks/app/api/(settings)/getEmployee/route-7ac9828e0cf7acb0.js",revision:"7ac9828e0cf7acb0"},{url:"/_next/static/chunks/app/api/(tasco)/getTascoLog/comment/%5BtimeSheetId%5D/route-db2514bbe36d609a.js",revision:"db2514bbe36d609a"},{url:"/_next/static/chunks/app/api/(tasco)/getTascoLog/loadCount/%5BtimeSheetId%5D/route-e1d7e348685bd47e.js",revision:"e1d7e348685bd47e"},{url:"/_next/static/chunks/app/api/(tasco)/getTascoLog/refueledLogs/%5BtimeSheetId%5D/route-048e1a7248d42c53.js",revision:"048e1a7248d42c53"},{url:"/_next/static/chunks/app/api/(tasco)/getTascoLog/tascoId/route-4c82b5f5bdce889d.js",revision:"4c82b5f5bdce889d"},{url:"/_next/static/chunks/app/api/(timeSheets)/getComment/route-000b35c9a7c0e1ec.js",revision:"000b35c9a7c0e1ec"},{url:"/_next/static/chunks/app/api/(timeSheets)/getMaterialTypes/route-f6d6864a813f60fb.js",revision:"f6d6864a813f60fb"},{url:"/_next/static/chunks/app/api/(trucking)/getTruckingLogs/endingMileage/%5BtimeSheetId%5D/route-5eafa7cfe7885ec8.js",revision:"5eafa7cfe7885ec8"},{url:"/_next/static/chunks/app/api/(trucking)/getTruckingLogs/equipmentHauled/%5BtimeSheetId%5D/route-56267620569b4d7f.js",revision:"56267620569b4d7f"},{url:"/_next/static/chunks/app/api/(trucking)/getTruckingLogs/material/%5BtimeSheetId%5D/route-77bd5bff7071b9f1.js",revision:"77bd5bff7071b9f1"},{url:"/_next/static/chunks/app/api/(trucking)/getTruckingLogs/notes/%5BtimeSheetId%5D/route-935cbb96f3d803f5.js",revision:"935cbb96f3d803f5"},{url:"/_next/static/chunks/app/api/(trucking)/getTruckingLogs/refueledLogs/%5BtimeSheetId%5D/route-e3a0e81daba2f047.js",revision:"e3a0e81daba2f047"},{url:"/_next/static/chunks/app/api/(trucking)/getTruckingLogs/startingMileage/%5BtimeSheetId%5D/route-8608c51eb55c9658.js",revision:"8608c51eb55c9658"},{url:"/_next/static/chunks/app/api/(trucking)/getTruckingLogs/stateMileage/%5BtimeSheetId%5D/route-8160ab7228d3e6e1.js",revision:"8160ab7228d3e6e1"},{url:"/_next/static/chunks/app/api/(trucking)/getTruckingLogs/truckingId/route-dd0edb856a340d50.js",revision:"dd0edb856a340d50"},{url:"/_next/static/chunks/app/api/(user)/getUserImage/route-6adf5a7de0093e39.js",revision:"6adf5a7de0093e39"},{url:"/_next/static/chunks/app/api/auth/%5B...nextauth%5D/route-6d1e12b682014910.js",revision:"6d1e12b682014910"},{url:"/_next/static/chunks/app/api/clockoutDetails/route-352ff9e51ca65180.js",revision:"352ff9e51ca65180"},{url:"/_next/static/chunks/app/api/equipment/%5BequipmentId%5D/lastMileage/route-94419db7d73a8237.js",revision:"94419db7d73a8237"},{url:"/_next/static/chunks/app/api/getTimesheetDetailsManager/%5Bid%5D/route-e01aade37f72fe2b.js",revision:"e01aade37f72fe2b"},{url:"/_next/static/chunks/app/api/getTimesheetsByDate/route-d3773ff605e31d2e.js",revision:"d3773ff605e31d2e"},{url:"/_next/static/chunks/app/api/getTimesheetsByDateNew/route-d1db4096be08a38b.js",revision:"d1db4096be08a38b"},{url:"/_next/static/chunks/app/api/getTruckingLogs/equipmentId/%5BtimeSheetId%5D/route-3e1bd301359b6d6a.js",revision:"3e1bd301359b6d6a"},{url:"/_next/static/chunks/app/api/sentry-example-api/route-3ad3b080b6289ca2.js",revision:"3ad3b080b6289ca2"},{url:"/_next/static/chunks/app/api/truckingLogDetails/%5BtruckingLogId%5D/route-f424dae265011067.js",revision:"f424dae265011067"},{url:"/_next/static/chunks/app/api/user-permissions/route-49bef76c98d83ef7.js",revision:"49bef76c98d83ef7"},{url:"/_next/static/chunks/app/error-8ed5a2d151f85834.js",revision:"8ed5a2d151f85834"},{url:"/_next/static/chunks/app/global-error-293467dba0cb08e6.js",revision:"293467dba0cb08e6"},{url:"/_next/static/chunks/app/icons/page-c8cf36ee1169cb53.js",revision:"c8cf36ee1169cb53"},{url:"/_next/static/chunks/app/layout-923910bbb72323dc.js",revision:"923910bbb72323dc"},{url:"/_next/static/chunks/app/not-authorized/page-1f20afeac966c607.js",revision:"1f20afeac966c607"},{url:"/_next/static/chunks/app/not-found-25775581529f7a18.js",revision:"25775581529f7a18"},{url:"/_next/static/chunks/app/page-3d0d29b8136fb1b4.js",revision:"3d0d29b8136fb1b4"},{url:"/_next/static/chunks/framework-5871c7b6f4022a8e.js",revision:"5871c7b6f4022a8e"},{url:"/_next/static/chunks/main-3c0f3bf73ac5d28c.js",revision:"3c0f3bf73ac5d28c"},{url:"/_next/static/chunks/main-app-dc8d286d71daa073.js",revision:"dc8d286d71daa073"},{url:"/_next/static/chunks/pages/_app-5376b67503103dac.js",revision:"5376b67503103dac"},{url:"/_next/static/chunks/pages/_error-e94a0482ccaf0c05.js",revision:"e94a0482ccaf0c05"},{url:"/_next/static/chunks/polyfills-42372ed130431b0a.js",revision:"846118c33b2c0e922d7b3a7676f81f6f"},{url:"/_next/static/chunks/webpack-7508aada89e7c79a.js",revision:"7508aada89e7c79a"},{url:"/_next/static/css/1b60633aeacf460b.css",revision:"1b60633aeacf460b"},{url:"/_next/static/css/63c7182515913e2f.css",revision:"63c7182515913e2f"},{url:"/_next/static/css/ae3f615d9a81d81a.css",revision:"ae3f615d9a81d81a"},{url:"/_next/static/css/b5fe2f5a95132839.css",revision:"b5fe2f5a95132839"},{url:"/_next/static/css/ccce47fdcedd91d2.css",revision:"ccce47fdcedd91d2"},{url:"/_next/static/media/30e6a981a0efeb9b-s.woff2",revision:"5cd14db2bb47ab5db74376b77c909759"},{url:"/_next/static/media/5df40a3eb8df1a4c-s.p.woff2",revision:"2f46a0d75dfff5f623a08402583e7ee9"},{url:"/_next/static/media/62d466ce3199db61-s.p.woff2",revision:"3a28d888908be7f560bf59faa1c796b2"},{url:"/_next/static/media/92c2336771404627-s.woff2",revision:"e456d9e276ba68ad76f65962e88fab81"},{url:"/_next/static/media/ajax-loader.0b80f665.gif",revision:"0b80f665"},{url:"/_next/static/media/d934919622a522a7-s.woff2",revision:"d3322001406183e84b58b99afd278b84"},{url:"/_next/static/media/slick.25572f22.eot",revision:"25572f22"},{url:"/_next/static/media/slick.653a4cbb.woff",revision:"653a4cbb"},{url:"/_next/static/media/slick.6aa1ee46.ttf",revision:"6aa1ee46"},{url:"/_next/static/media/slick.f895cfdf.svg",revision:"f895cfdf"},{url:"/admin-white.svg",revision:"f60bbe82328c01c11b8e0b2970c889e6"},{url:"/admin.svg",revision:"0d220493364872b69d6f6cde6e385d2c"},{url:"/android/android-launchericon-144-144.png",revision:"554cb47f9b447d7c049e0af337f30598"},{url:"/android/android-launchericon-192-192.png",revision:"c7acce3a1ca251dc273817cd5462a666"},{url:"/android/android-launchericon-48-48.png",revision:"b3e487bbda706b0a7a31bd383fcb5bb9"},{url:"/android/android-launchericon-512-512.png",revision:"daaca9c8424139dcfdeed3181b0a7cbd"},{url:"/android/android-launchericon-72-72.png",revision:"0544816fbb562d0825feb6e9ff73edff"},{url:"/android/android-launchericon-96-96.png",revision:"29c67668a313466cbd66948974b9c656"},{url:"/arrowBack.svg",revision:"17d544aac2962f3f0b83d7dd8bf8d5e7"},{url:"/arrowDown.svg",revision:"bde6e8ea1fa48902e773fc0f34d8ad2c"},{url:"/arrowDownThin.svg",revision:"c8dc31e8dcdca0b4feb827ad7b28c336"},{url:"/arrowLeft.svg",revision:"548bd36e974c05627ee6042a0d17e6c6"},{url:"/arrowLeftSymbol.svg",revision:"4cfdc72b6b5f2cf11ea191dc60c5c4b4"},{url:"/arrowLeftThin.svg",revision:"c4bbdff0ab65d617688561fee993f1c1"},{url:"/arrowRight.svg",revision:"36ae3b90062b7432459cad98be741a80"},{url:"/arrowRightSymbol.svg",revision:"e771082a8a75d50f59df162c67d26091"},{url:"/arrowRightThin.svg",revision:"73a5a9d7e4bf4951d11c4e268faeee85"},{url:"/arrowUp.svg",revision:"3e3a91581069f2b5dc59c7c465255263"},{url:"/arrowUpThin.svg",revision:"ef72218a70f19849149176593f9a8cee"},{url:"/biometrics.svg",revision:"f42bc2055fc95544b3641a2819a084d7"},{url:"/broken.svg",revision:"c62a555a912a5887033a4661236ae31d"},{url:"/calendar.svg",revision:"d6aa1ba02d5a6253fcec8da68242da2a"},{url:"/camera.svg",revision:"4657cc3ac5e2a543f0829e6100446591"},{url:"/cameraFilled.svg",revision:"bcdbce7deaeb9a1d37058d3befea92fc"},{url:"/checkbox.svg",revision:"c2c40521ced6fc4ba77a3865d9dc345f"},{url:"/clock.svg",revision:"eff35cd79029bdd0eeebc39b052f0fbf"},{url:"/clockBreak.svg",revision:"0f0d1da56fbbed48b97b5dbba7d34e65"},{url:"/clockIn.svg",revision:"02e58b9c6f4a129a7026400944bd15b7"},{url:"/clockOut.svg",revision:"a8b9f471daf782532b4006c2cb20b89a"},{url:"/comment.svg",revision:"7bb5715a71c021149b5c3b811aac572b"},{url:"/comments.svg",revision:"29b1a7eb27c708878a454c2eeaad616d"},{url:"/condense-white.svg",revision:"6e9075aa9d7fcb1c242b78610ea9330a"},{url:"/condense.svg",revision:"bc4d6b17d44724df96d6d4a3016072da"},{url:"/drag.svg",revision:"da36f043748f4e8af0e0f5ee17e289a2"},{url:"/dragDots.svg",revision:"c9d25bda626917ea72ca49977f11a016"},{url:"/dragFormBuilder.svg",revision:"bfdedbf1bc6d64afb3642848a542f55d"},{url:"/endDay.svg",revision:"78010192db1eaa29e95ac4d2d9f3705e"},{url:"/equipment-white.svg",revision:"0760c1e58029025d443832cdc33d7402"},{url:"/equipment.svg",revision:"eeb87295d51e91e2adb6483c18bc2452"},{url:"/eraser.svg",revision:"7ed7c543a0ff5530d3c586b368c51315"},{url:"/export-white.svg",revision:"4793d42ec50e11a8586bbc2a5a2f59c8"},{url:"/export.svg",revision:"09f9524501c9e871487735c2b09fcc43"},{url:"/eye-blue.svg",revision:"a851e7f5c035279390dfa1eaa3387e0d"},{url:"/eye.svg",revision:"b7243fcb645ec88a90d9b66fdc2abca3"},{url:"/eyeSlash.svg",revision:"3bb6de382b51b3a333e146bcf8a5d4f5"},{url:"/favicon_io/android-chrome-192x192.png",revision:"07281900d7ffa5690cc984da6e613805"},{url:"/favicon_io/android-chrome-512x512.png",revision:"ee520bf09fd4e101b8952887694905fc"},{url:"/favicon_io/apple-touch-icon.png",revision:"6f682a595255e6aab2bf0f2daa78403a"},{url:"/favicon_io/favicon-16x16.png",revision:"2d00862eea3e0a26607d3d0ce24d1747"},{url:"/favicon_io/favicon-32x32.png",revision:"84f7a4563c6fb8cd905f2dbedcffede8"},{url:"/fileClosed.svg",revision:"bf5a8aeebd62c1736f9e0062a3a271a0"},{url:"/fileOpen.svg",revision:"9ce1ceadae4b4abba237029e0a410d8f"},{url:"/filterDials.svg",revision:"13341b5339e076be7b361c2e09e4976c"},{url:"/filterFunnel.svg",revision:"743604a4e2e50abbb135746948880f65"},{url:"/form-white.svg",revision:"5f85274db63c253161739424aa2fab34"},{url:"/form.svg",revision:"5828ab7b08b01307092a48136ac39cb0"},{url:"/formApproval.svg",revision:"a53cba1fffa0cf55e4dbd86d0345b5e8"},{url:"/formDuplicate.svg",revision:"898bd36ec2c706978dae4e684575ec3e"},{url:"/formEdit.svg",revision:"017c59837c434df47dce4b457e71996e"},{url:"/formInspect-white.svg",revision:"8e48a64eb645a8b5bdc204706e8e6451"},{url:"/formInspect.svg",revision:"f2fcd14456afc03a583c568f4c4da3c8"},{url:"/formList-white.svg",revision:"e1ced4954a119850e3f06e0980062497"},{url:"/formList.svg",revision:"aef95c6f9019999cb9992dd2a21b3975"},{url:"/formSave.svg",revision:"e504c19b5bbbe70b0d2c83c9a73927b8"},{url:"/formSent.svg",revision:"eca064a2079c76fd61ca20528716ceee"},{url:"/formUndo.svg",revision:"91f221fce7b16f414253f9f1873902b2"},{url:"/hauling.svg",revision:"57a718f3ba6755d80106a01a21063d79"},{url:"/haulingFilled.svg",revision:"ebe07b7f9ca47cc45f7fd6b48e6d02ca"},{url:"/header.svg",revision:"479adabc8b70a52625854fba7e433548"},{url:"/home-white.svg",revision:"50f6d70fd4e0a491044340d69138b394"},{url:"/home.svg",revision:"c78e854f7b56bb952a0a22fc1980bc7d"},{url:"/icon-192X192.png",revision:"47aa596be976071dfff9c1e5455ecc80"},{url:"/icon-384x384.png",revision:"99d1987d20b874836d950b95586968ae"},{url:"/icon-512x512.png",revision:"c16feee422d814c946012979c0d02f00"},{url:"/inbox-white.svg",revision:"dfb451099519b395ed3d9b15daad5708"},{url:"/inbox.svg",revision:"03332320e762076df175dde846d62022"},{url:"/inboxFilled.svg",revision:"8b5e6d06de80a9927ae85007277b98a9"},{url:"/information.svg",revision:"3e9a1c81c8fd396fdf6849975d62dbb5"},{url:"/injury.svg",revision:"18121c3e1dc867f5d0a332a7133057f4"},{url:"/ios/100.png",revision:"5222167951b54f250720976e5798bd81"},{url:"/ios/1024.png",revision:"4046bc171d373187ec11f28c06ceb334"},{url:"/ios/114.png",revision:"2aec0491cbf7eb22d34f3079cf2eff54"},{url:"/ios/120.png",revision:"52ac3d673f239d4c822b20c337c73db2"},{url:"/ios/128.png",revision:"bc9dd541b1a0aaff432df4f5b0fb88b4"},{url:"/ios/144.png",revision:"554cb47f9b447d7c049e0af337f30598"},{url:"/ios/152.png",revision:"3085835ae98a412b5f8bb41b3d7fccae"},{url:"/ios/16.png",revision:"876eb7a741ae46958961bf0ed6c30ba0"},{url:"/ios/167.png",revision:"a7230785a2891b4659792d7d6cab6ffb"},{url:"/ios/180.png",revision:"5cdb37c8967955e3a4fae232138f3c03"},{url:"/ios/192.png",revision:"c7acce3a1ca251dc273817cd5462a666"},{url:"/ios/20.png",revision:"8c0f62f1ddd74accb89e318fe432a70f"},{url:"/ios/256.png",revision:"19665fd60a705b586e5cc5c93eddc1f2"},{url:"/ios/29.png",revision:"536c6f0fcf359bba5e0b45c9bb4fc260"},{url:"/ios/32.png",revision:"55a60f12c21ea3fffa61d8f7a34cfc33"},{url:"/ios/40.png",revision:"a38e8ea2895f2cfe54be4ce7465e7952"},{url:"/ios/50.png",revision:"db849275edb671b9ea7c45ef6f4fc489"},{url:"/ios/512.png",revision:"daaca9c8424139dcfdeed3181b0a7cbd"},{url:"/ios/57.png",revision:"8511d6a6a25878417899d2041de166ff"},{url:"/ios/58.png",revision:"001948872d9d885d20140b7aa9cc6ff8"},{url:"/ios/60.png",revision:"80209b85fb0be10562bcee31908d2a95"},{url:"/ios/64.png",revision:"066fa2094719e23427225f5517889ed7"},{url:"/ios/72.png",revision:"0544816fbb562d0825feb6e9ff73edff"},{url:"/ios/76.png",revision:"0a1c5da7e5a9673345c44f7e1c87f0d4"},{url:"/ios/80.png",revision:"91b74cbd7d0f0119a164a0750462149e"},{url:"/ios/87.png",revision:"74c6a6624174372f1fc483515c820c2f"},{url:"/jobsite-white.svg",revision:"1d991bcf439a2538ed78c709c708954b"},{url:"/jobsite.svg",revision:"f3909b4ff16c4f25e211fb0dc7556050"},{url:"/key.svg",revision:"f39b7da7fb64da838435e228ca3113ed"},{url:"/language.svg",revision:"e4615572d416d5df80c463dc53d5e0b7"},{url:"/layout.svg",revision:"a3b50fb72e8fab8feac60ccbd864e31a"},{url:"/logo.svg",revision:"8e78aff787124cc6567c2ccfd520099d"},{url:"/logo_JPG.jpg",revision:"e84c38cbd142e140688beacd59e6a0cf"},{url:"/manifest.json",revision:"05d4e6a0d9c7c7bebec473d08ea582b2"},{url:"/mechanic.svg",revision:"4f39d9d1ff5c68a81b93c7f6611d8e9d"},{url:"/message.svg",revision:"c55f300ce914f4b28036e30580ee104c"},{url:"/mileage.svg",revision:"11e1eb1d01c9be54e3c8394e198db717"},{url:"/minus.svg",revision:"d2ec83afd3490cac9da747386df54184"},{url:"/moreOptions.svg",revision:"d3019cfec5d1102ec73147f1dd5e5faf"},{url:"/moreOptionsCircle.svg",revision:"02d86b2a2acd7707b4cef84da6ede884"},{url:"/number.svg",revision:"6f90bfd3f431fe8602537fe47b9a109b"},{url:"/plus-stroke-white.svg",revision:"f23aeed115ad4273e185d3e73a314ba7"},{url:"/plus-white.svg",revision:"a3d859abbf1ce862eb0aa1d070d85fbf"},{url:"/plus.svg",revision:"33d3fcf5251a72f1eef80aaa19f7ab55"},{url:"/policies.svg",revision:"25de43083a5e24cdf83a7ba62d6c28c4"},{url:"/priorityDelay.svg",revision:"5286ef619f78a44968b600b83bba55a0"},{url:"/priorityHigh.svg",revision:"fa44198aecf39d4748914d7f3b4a0d03"},{url:"/priorityLow.svg",revision:"11efcad98767f7c049d90ece2f425f1a"},{url:"/priorityMedium.svg",revision:"219c8a19b23f94db24162d544e1605ec"},{url:"/priorityPending.svg",revision:"45d3ef273daf2b4daadcb0b6d0064789"},{url:"/priorityToday.svg",revision:"91fec1dfbc8a1268488c12a9ef9fa4ac"},{url:"/profileEmpty.svg",revision:"66491e65dd6114a9acf6990181fc44d9"},{url:"/profileFilled.svg",revision:"ff2e851ef303666374501b541725d34c"},{url:"/qrCode-white.svg",revision:"0bcc19567973489fa8a15e759b729d5b"},{url:"/qrCode.svg",revision:"a97e34eddfd5bae153771e734fd96e4c"},{url:"/radio.svg",revision:"17ecc9b4e325da65fc947c71d4bc1ade"},{url:"/refuel.svg",revision:"07fcabd768272b84f61007c4e130b8fc"},{url:"/refuelFilled.svg",revision:"d75e2ee23202812499b36b050fff841c"},{url:"/searchLeft.svg",revision:"7a4c52a1de03e0fbe1f2f949fed97367"},{url:"/searchRight.svg",revision:"9ac605dce8646197fa3ae4c90d93f8c9"},{url:"/settingsFilled.svg",revision:"4b9876c68beb7435b569e5a5a0252501"},{url:"/shiftScanLogo.svg",revision:"f6cdbf4c760167995264751f68ebb2af"},{url:"/shiftscanlogoHorizontal.svg",revision:"f6dac5483d88c1cdebc5f92190cc1d2f"},{url:"/spinner.svg",revision:"7eedb8067da00e29baf517b3cbfffc43"},{url:"/star.svg",revision:"744181e34e34c3385a845fa37cf51454"},{url:"/starFilled.svg",revision:"5ad7cd14d1deca592924c230b5a39218"},{url:"/state.svg",revision:"83c029784c461361996615f83c3850ef"},{url:"/stateFilled.svg",revision:"183c533850d5a2540b71b2c44bef9d11"},{url:"/statusApproved.svg",revision:"0e767f6c9f2c8f20480d3120d6b6e198"},{url:"/statusApprovedFilled.svg",revision:"6fc76b4c67d20fa9ffbdb451a30e8071"},{url:"/statusDenied.svg",revision:"7b1167c6857f8169396ae88ffbce95e6"},{url:"/statusDeniedFilled.svg",revision:"52740140811df8fc4cd87604621c9960"},{url:"/statusOffline.svg",revision:"d1931f81ba9fe9342e78e181c2216aaa"},{url:"/statusOngoing-white.svg",revision:"d07cb3a9608fbfa3e64486e49eb74365"},{url:"/statusOngoing.svg",revision:"7eedb8067da00e29baf517b3cbfffc43"},{url:"/statusOngoingFilled.svg",revision:"5e18aea67f178cddbdc7e2e449ea6e3a"},{url:"/statusOnline.svg",revision:"ebed6ae0692535bcfee10ada59c779c3"},{url:"/statusUnfinished.svg",revision:"038f7353e12e76758c9e214925a07a0c"},{url:"/swe-worker-5c72df51bb1f6ee0.js",revision:"447aee4a2c24fefe7917563f0255d688"},{url:"/tasco.svg",revision:"1280d07dd40b7604d022f71c3728ccfb"},{url:"/team-white.svg",revision:"0142ac2317a2b44b141c270916ab32ca"},{url:"/team.svg",revision:"203a48a7519f73fe7f24fff1cae20204"},{url:"/timecards.svg",revision:"7771f6f39b0e3d775e88c6d542d2d68b"},{url:"/tinyCheckMark-white.svg",revision:"a6d6eadc3dbfaafbf1e8257346eee02c"},{url:"/title.svg",revision:"15b1291ddf60f0ceb36ce46377fa6dd3"},{url:"/trash-red.svg",revision:"a42dfceb5b388b58d0c41c13ee0c17f7"},{url:"/trash.svg",revision:"229905df8ec07dd400f7a23deb582f0c"},{url:"/trucking.svg",revision:"ab3794a1a3073bc2926a1978a038300c"},{url:"/user-white.svg",revision:"6f6e9f9b948a6dfde98ca0b41985bbc0"},{url:"/user.svg",revision:"8756e7b0a6bc7ba8068b0f0806b5e5e7"},{url:"/windows11/LargeTile.scale-100.png",revision:"e3ae3637973ea93ff81e73ee37effe0c"},{url:"/windows11/LargeTile.scale-125.png",revision:"6e4ec698528b71f25efd6ff2dd68e89f"},{url:"/windows11/LargeTile.scale-150.png",revision:"b636fa1812d337a4ca1224b433a3d855"},{url:"/windows11/LargeTile.scale-200.png",revision:"d18a045105d03ba23b624c109708f7bb"},{url:"/windows11/LargeTile.scale-400.png",revision:"d4176f9ebaacc0bbfebf56fd56e253ce"},{url:"/windows11/SmallTile.scale-100.png",revision:"bce1445e4fb91c8c0842780a4026522e"},{url:"/windows11/SmallTile.scale-125.png",revision:"9415b7ff287961c074dee1834b42aad9"},{url:"/windows11/SmallTile.scale-150.png",revision:"87d19729c39b8d355a62d180d91bb31b"},{url:"/windows11/SmallTile.scale-200.png",revision:"7c5db43e6077ca3c958c21c85bec6d25"},{url:"/windows11/SmallTile.scale-400.png",revision:"1913d6ed56bc5b31dc60858acbcace6f"},{url:"/windows11/SplashScreen.scale-100.png",revision:"a8215f24409e6231d1dc00a0907ed4a8"},{url:"/windows11/SplashScreen.scale-125.png",revision:"097ac942f0459d3481fbd62c0f97e299"},{url:"/windows11/SplashScreen.scale-150.png",revision:"201fc53647b740af27d9b004cafeaaf8"},{url:"/windows11/SplashScreen.scale-200.png",revision:"1b1081f94203f3823ed2dd1f8dabb29a"},{url:"/windows11/SplashScreen.scale-400.png",revision:"0d6ee6f415a7d2101792e2b8312af059"},{url:"/windows11/Square150x150Logo.scale-100.png",revision:"7b6855c0928dad5db20403341b059afc"},{url:"/windows11/Square150x150Logo.scale-125.png",revision:"dd8460f97de8d2cf5826468f8f8153db"},{url:"/windows11/Square150x150Logo.scale-150.png",revision:"8bfd4727bbeb856bd1af17e0cdef44ea"},{url:"/windows11/Square150x150Logo.scale-200.png",revision:"c95f1e579117442bcf6a388818260303"},{url:"/windows11/Square150x150Logo.scale-400.png",revision:"a20add32148ad6bfcc0d435ac4986eb5"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-16.png",revision:"90a9d6c44b09d44f416b62df26224428"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-20.png",revision:"779afd7d8c738f0a0c73a4a6d63e4aff"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-24.png",revision:"a65f2b674acb5c2af36db18e9b4a63f1"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-256.png",revision:"723b833ab63af25543dfe92f0a56e6d5"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-30.png",revision:"3a63be52b054d426a0ee4ce69b783e12"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-32.png",revision:"261d9c7ca92fefc2230721bb99cbcae1"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-36.png",revision:"a709411193e6374836f5fa82b8afa6c1"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-40.png",revision:"4251a65706de0f7c86b1bee205eb53a7"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-44.png",revision:"9f4cf9e0ba353f1bbb19fc9f6e40523a"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-48.png",revision:"c87a2b33f89afc0199219ea0b6e191ac"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-60.png",revision:"727adb65708020078e70655ef5444a7b"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-64.png",revision:"04915cfcb9f9ad29a6223fb193c79a14"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-72.png",revision:"1f5e7f8fb46e7d8e425ff06c34338b6d"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-80.png",revision:"3a7aa9e5f873375880d85aa0c2f51306"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-96.png",revision:"a151f9aefbcedac0d11aba5cf1108af8"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-16.png",revision:"90a9d6c44b09d44f416b62df26224428"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-20.png",revision:"779afd7d8c738f0a0c73a4a6d63e4aff"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-24.png",revision:"a65f2b674acb5c2af36db18e9b4a63f1"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-256.png",revision:"723b833ab63af25543dfe92f0a56e6d5"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-30.png",revision:"3a63be52b054d426a0ee4ce69b783e12"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-32.png",revision:"261d9c7ca92fefc2230721bb99cbcae1"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-36.png",revision:"a709411193e6374836f5fa82b8afa6c1"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-40.png",revision:"4251a65706de0f7c86b1bee205eb53a7"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-44.png",revision:"9f4cf9e0ba353f1bbb19fc9f6e40523a"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-48.png",revision:"c87a2b33f89afc0199219ea0b6e191ac"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-60.png",revision:"727adb65708020078e70655ef5444a7b"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-64.png",revision:"04915cfcb9f9ad29a6223fb193c79a14"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-72.png",revision:"1f5e7f8fb46e7d8e425ff06c34338b6d"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-80.png",revision:"3a7aa9e5f873375880d85aa0c2f51306"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-96.png",revision:"a151f9aefbcedac0d11aba5cf1108af8"},{url:"/windows11/Square44x44Logo.scale-100.png",revision:"9f4cf9e0ba353f1bbb19fc9f6e40523a"},{url:"/windows11/Square44x44Logo.scale-125.png",revision:"95498060191b331348eb8f5b5ab6dfe7"},{url:"/windows11/Square44x44Logo.scale-150.png",revision:"aced07a7a4627cdad487ed0a46faeebb"},{url:"/windows11/Square44x44Logo.scale-200.png",revision:"db11fa1c0b22e2344ecf9dea9af68a2c"},{url:"/windows11/Square44x44Logo.scale-400.png",revision:"5f1b23fb222c880bf3c533643f43b4ca"},{url:"/windows11/Square44x44Logo.targetsize-16.png",revision:"90a9d6c44b09d44f416b62df26224428"},{url:"/windows11/Square44x44Logo.targetsize-20.png",revision:"779afd7d8c738f0a0c73a4a6d63e4aff"},{url:"/windows11/Square44x44Logo.targetsize-24.png",revision:"a65f2b674acb5c2af36db18e9b4a63f1"},{url:"/windows11/Square44x44Logo.targetsize-256.png",revision:"723b833ab63af25543dfe92f0a56e6d5"},{url:"/windows11/Square44x44Logo.targetsize-30.png",revision:"3a63be52b054d426a0ee4ce69b783e12"},{url:"/windows11/Square44x44Logo.targetsize-32.png",revision:"261d9c7ca92fefc2230721bb99cbcae1"},{url:"/windows11/Square44x44Logo.targetsize-36.png",revision:"a709411193e6374836f5fa82b8afa6c1"},{url:"/windows11/Square44x44Logo.targetsize-40.png",revision:"4251a65706de0f7c86b1bee205eb53a7"},{url:"/windows11/Square44x44Logo.targetsize-44.png",revision:"9f4cf9e0ba353f1bbb19fc9f6e40523a"},{url:"/windows11/Square44x44Logo.targetsize-48.png",revision:"c87a2b33f89afc0199219ea0b6e191ac"},{url:"/windows11/Square44x44Logo.targetsize-60.png",revision:"727adb65708020078e70655ef5444a7b"},{url:"/windows11/Square44x44Logo.targetsize-64.png",revision:"04915cfcb9f9ad29a6223fb193c79a14"},{url:"/windows11/Square44x44Logo.targetsize-72.png",revision:"1f5e7f8fb46e7d8e425ff06c34338b6d"},{url:"/windows11/Square44x44Logo.targetsize-80.png",revision:"3a7aa9e5f873375880d85aa0c2f51306"},{url:"/windows11/Square44x44Logo.targetsize-96.png",revision:"a151f9aefbcedac0d11aba5cf1108af8"},{url:"/windows11/StoreLogo.scale-100.png",revision:"db849275edb671b9ea7c45ef6f4fc489"},{url:"/windows11/StoreLogo.scale-125.png",revision:"728e6527e2d77371e3007caf007b116f"},{url:"/windows11/StoreLogo.scale-150.png",revision:"9224f9a45257c2103e10aa9026af46d1"},{url:"/windows11/StoreLogo.scale-200.png",revision:"5222167951b54f250720976e5798bd81"},{url:"/windows11/StoreLogo.scale-400.png",revision:"6562c92cea67a6232c5727e464a9dd83"},{url:"/windows11/Wide310x150Logo.scale-100.png",revision:"7fc14247c4af58216f16d5f4f98fc46b"},{url:"/windows11/Wide310x150Logo.scale-125.png",revision:"a0c8d4ccacb2bc65438458df95ab3d60"},{url:"/windows11/Wide310x150Logo.scale-150.png",revision:"34161357c0cdafe2651b8d197c019798"},{url:"/windows11/Wide310x150Logo.scale-200.png",revision:"a8215f24409e6231d1dc00a0907ed4a8"},{url:"/windows11/Wide310x150Logo.scale-400.png",revision:"1b1081f94203f3823ed2dd1f8dabb29a"}],{ignoreURLParametersMatching:[/^utm_/,/^fbclid$/]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({response:e})=>e&&"opaqueredirect"===e.type?new Response(e.body,{status:200,statusText:"OK",headers:e.headers}):e}]}),"GET"),e.registerRoute(/^https?.*/,new e.NetworkFirst({cacheName:"offlineCache",plugins:[new e.ExpirationPlugin({maxEntries:200,maxAgeSeconds:86400})]}),"GET"),self.__WB_DISABLE_DEV_LOGS=!0}));
 //# sourceMappingURL=sw.js.map
