+++ shift-scan/src/app/(routes)/admins/equipment/_components/useEquipmentData.ts
@@ -1,5 +1,8 @@
 "use client";
 import { useState, useEffect, useCallback } from "react";
+import QRCode from "qrcode";
+import { deleteEquipment } from "@/actions/AssetActions";
+import { useSidebar } from "@/components/ui/sidebar";
 
 /**
  * EquipmentVehicleInfo type for vehicle-specific info
@@ -29,6 +32,7 @@ export const useEquipmentData = () => {
   const [equipmentDetails, setEquipmentDetails] = useState<EquipmentSummary[]>(
     [],
   );
+  const [searchTerm, setSearchTerm] = useState("");
   const [loading, setLoading] = useState<boolean>(true);
   const [refreshKey, setRefreshKey] = useState(0);
   const [total, setTotal] = useState<number>(0);
@@ -36,6 +40,13 @@ export const useEquipmentData = () => {
   const [pageSize, setPageSize] = useState<number>(25);
   const [totalPages, setTotalPages] = useState<number>(0);
   const [pendingCount, setPendingCount] = useState<number>(0);
+  // State for modals
+  const [editEquipmentModal, setEditEquipmentModal] = useState(false);
+  const [createEquipmentModal, setCreateEquipmentModal] = useState(false);
+  const [showDeleteDialog, setShowDeleteDialog] = useState(false);
+  const [pendingDeleteId, setPendingDeleteId] = useState<string | null>(null);
+  const [pendingEditId, setPendingEditId] = useState<string | null>(null);
+  const [pendingQrId, setPendingQrId] = useState<string | null>(null);
 
   //Approval Button States
   const [showPendingOnly, setShowPendingOnly] = useState(false);
@@ -74,25 +85,175 @@ export const useEquipmentData = () => {
     fetchEquipmentSummaries();
   }, [refreshKey, page, pageSize, showPendingOnly]);
 
+  const openHandleEdit = (id: string) => {
+    setPendingEditId(id);
+    setEditEquipmentModal(true);
+  };
+
+  const openHandleDelete = (id: string) => {
+    setPendingDeleteId(id);
+    setShowDeleteDialog(true);
+  };
+
+  const confirmDelete = async () => {
+    if (pendingDeleteId) {
+      await deleteEquipment(pendingDeleteId);
+      setShowDeleteDialog(false);
+      setPendingDeleteId(null);
+      rerender();
+    }
+  };
+
+  useEffect(() => {
+    setPage(1);
+  }, [searchTerm, showPendingOnly]);
+
+  const openHandleQr = (id: string) => {
+    console.log("openHandleQr called with id:", id);
+    setPendingQrId(id);
+    const equipment = equipmentDetails.find((j) => j.id === id);
+    if (equipment) {
+      printQRCode(equipment);
+    }
+  };
+
+  const printQRCode = async (equipment: EquipmentSummary) => {
+    if (!pendingQrId) return;
+    const url = await QRCode.toDataURL(equipment.qrId || "");
+    // Open a new window for printing
+    const printWindow = window.open("", "_blank");
+    if (!printWindow) {
+      alert("Please allow popups to print the QR code");
+      return;
+    }
+
+    // Write HTML content to the new window
+    printWindow.document.write(`
+      <!DOCTYPE html>
+      <html>
+      <head>
+        <title>Print QR Code - ${equipment.name || "Equipment"}</title>
+        <style>
+          body {
+            display: flex;
+            flex-direction: column;
+            align-items: center;
+            justify-content: center;
+            height: 100vh;
+            margin: 0;
+            padding: 20px;
+            box-sizing: border-box;
+            font-family: Arial, sans-serif;
+          }
+          .qr-code-container {
+            text-align: center;
+          }
+          .qr-code {
+            width: 300px;
+            height: 300px;
+            border: 4px solid black;
+            border-radius: 10px;
+            margin-bottom: 20px;
+          }
+          .equipment-name {
+            font-size: 24px;
+            font-weight: bold;
+            margin-bottom: 10px;
+          }
+          .equipment-id {
+            font-size: 16px;
+            color: #555;
+            margin-bottom: 8px;
+          }
+          .equipment-description {
+            font-size: 16px;
+            color: #555;
+            max-width: 350px;
+            padding: 0 20px;
+            line-height: 1.4;
+            margin-top: 8px;
+            white-space: pre-wrap;
+            overflow-wrap: break-word;
+          }
+          @media print {
+            body {
+              -webkit-print-color-adjust: exact;
+              color-adjust: exact;
+            }
+          }
+        </style>
+      </head>
+      <body>
+        <div class="qr-code-container">
+          <div class="equipment-name">${equipment.name || "N/A"}</div>
+          <img src="${url}" alt="QR Code" class="qr-code" />
+          <div class="equipment-id">ID: ${equipment.qrId || "N/A"}</div>
+          <div class="equipment-description">${
+            equipment.description
+              ? `Brief Description:\n${equipment.description || ""}`
+              : ""
+          }</div>
+        </div>
+        <script>
+          // Print and close window when loaded
+          window.onload = function() {
+            window.print();
+            // Close after printing is done or canceled
+            setTimeout(() => window.close());
+          };
+        </script>
+      </body>
+      </html>
+    `);
+
+    printWindow.document.close();
+  };
+
+  const cancelDelete = () => {
+    setShowDeleteDialog(false);
+    setPendingDeleteId(null);
+  };
+
+  // Filter equipment by name, make, or model, and by approval status if showPendingOnly is active
+  const filteredEquipment = equipmentDetails.filter((item) => {
+    const term = searchTerm.trim().toLowerCase();
+    if (!term) return true;
+    const nameMatch = item.name.toLowerCase().includes(term);
+    const makeMatch =
+      item.equipmentVehicleInfo?.make?.toLowerCase().includes(term) ?? false;
+    const modelMatch =
+      item.equipmentVehicleInfo?.model?.toLowerCase().includes(term) ?? false;
+    return nameMatch || makeMatch || modelMatch;
+  });
+
   return {
-    equipmentDetails,
     setEquipmentDetails,
     loading,
     setLoading,
     rerender,
-    // Pagination state
     total,
     page,
     pageSize,
     totalPages,
-    // Pagination handlers
-    setTotal,
     setPage,
     setPageSize,
-    setTotalPages,
-    // Approval Button States
     showPendingOnly,
     setShowPendingOnly,
     pendingCount,
+    editEquipmentModal,
+    setEditEquipmentModal,
+    createEquipmentModal,
+    setCreateEquipmentModal,
+    showDeleteDialog,
+    setShowDeleteDialog,
+    pendingEditId,
+    openHandleEdit,
+    openHandleDelete,
+    confirmDelete,
+    openHandleQr,
+    cancelDelete,
+    filteredEquipment,
+    searchTerm,
+    setSearchTerm,
   };
 };
