+++ shift-scan/src/app/(routes)/admins/equipment/page.tsx
@@ -1,14 +1,8 @@
 "use client";
 import { Button } from "@/components/ui/button";
 import { ScrollArea, ScrollBar } from "@/components/ui/scroll-area";
-import { useSidebar } from "@/components/ui/sidebar";
-import { useEffect, useState } from "react";
 import EquipmentTable from "./_components/equipmentTable";
-import {
-  EquipmentSummary,
-  useEquipmentData,
-} from "./_components/useEquipmentData";
-import QRCode from "qrcode";
+import { useEquipmentData } from "./_components/useEquipmentData";
 import {
   Dialog,
   DialogContent,
@@ -17,7 +11,6 @@ import {
   DialogHeader,
   DialogTitle,
 } from "@/components/ui/dialog";
-import { deleteEquipment } from "@/actions/AssetActions";
 import EditEquipmentModal from "./_components/EditEquipmentModal";
 import CreateEquipmentModal from "./_components/CreateEquipmentModal";
 import { Badge } from "@/components/ui/badge";
@@ -29,7 +22,6 @@ import {
   PaginationPrevious,
 } from "@/components/ui/pagination";
 import Spinner from "@/components/(animations)/spinner";
-import ReloadBtnSpinner from "@/components/(animations)/reload-btn-spinner";
 import SearchBarPopover from "../_pages/searchBarPopover";
 import {
   Tooltip,
@@ -39,176 +31,37 @@ import {
 import { PageHeaderContainer } from "../_pages/PageHeaderContainer";
 
 export default function EquipmentPage() {
-  const { setOpen, open } = useSidebar();
-  const [searchTerm, setSearchTerm] = useState("");
   const {
     loading,
-    equipmentDetails,
     rerender,
     total,
     page,
     pageSize,
     totalPages,
-    setTotal,
     setPage,
     setPageSize,
-    setTotalPages,
     showPendingOnly,
     setShowPendingOnly,
     pendingCount,
+    editEquipmentModal,
+    setEditEquipmentModal,
+    createEquipmentModal,
+    setCreateEquipmentModal,
+    showDeleteDialog,
+    setShowDeleteDialog,
+    pendingEditId,
+    openHandleEdit,
+    openHandleDelete,
+    confirmDelete,
+    openHandleQr,
+    cancelDelete,
+    filteredEquipment,
+    searchTerm,
+    setSearchTerm,
   } = useEquipmentData();
 
-  // State for modals
-  const [editEquipmentModal, setEditEquipmentModal] = useState(false);
-  const [createEquipmentModal, setCreateEquipmentModal] = useState(false);
-  const [showDeleteDialog, setShowDeleteDialog] = useState(false);
-  const [pendingDeleteId, setPendingDeleteId] = useState<string | null>(null);
-  const [pendingEditId, setPendingEditId] = useState<string | null>(null);
-  const [pendingQrId, setPendingQrId] = useState<string | null>(null);
-
-  const openHandleEdit = (id: string) => {
-    setPendingEditId(id);
-    setEditEquipmentModal(true);
-  };
-
-  const openHandleDelete = (id: string) => {
-    setPendingDeleteId(id);
-    setShowDeleteDialog(true);
-  };
-
-  const confirmDelete = async () => {
-    if (pendingDeleteId) {
-      await deleteEquipment(pendingDeleteId);
-      setShowDeleteDialog(false);
-      setPendingDeleteId(null);
-      rerender();
-    }
-  };
-
-  useEffect(() => {
-    setPage(1);
-  }, [searchTerm, showPendingOnly]);
-
-  const openHandleQr = (id: string) => {
-    console.log("openHandleQr called with id:", id);
-    setPendingQrId(id);
-    const equipment = equipmentDetails.find((j) => j.id === id);
-    if (equipment) {
-      printQRCode(equipment);
-    }
-  };
-
-  const printQRCode = async (equipment: EquipmentSummary) => {
-    if (!pendingQrId) return;
-    const url = await QRCode.toDataURL(equipment.qrId || "");
-    // Open a new window for printing
-    const printWindow = window.open("", "_blank");
-    if (!printWindow) {
-      alert("Please allow popups to print the QR code");
-      return;
-    }
-
-    // Write HTML content to the new window
-    printWindow.document.write(`
-      <!DOCTYPE html>
-      <html>
-      <head>
-        <title>Print QR Code - ${equipment.name || "Equipment"}</title>
-        <style>
-          body {
-            display: flex;
-            flex-direction: column;
-            align-items: center;
-            justify-content: center;
-            height: 100vh;
-            margin: 0;
-            padding: 20px;
-            box-sizing: border-box;
-            font-family: Arial, sans-serif;
-          }
-          .qr-code-container {
-            text-align: center;
-          }
-          .qr-code {
-            width: 300px;
-            height: 300px;
-            border: 4px solid black;
-            border-radius: 10px;
-            margin-bottom: 20px;
-          }
-          .equipment-name {
-            font-size: 24px;
-            font-weight: bold;
-            margin-bottom: 10px;
-          }
-          .equipment-id {
-            font-size: 16px;
-            color: #555;
-            margin-bottom: 8px;
-          }
-          .equipment-description {
-            font-size: 16px;
-            color: #555;
-            max-width: 350px;
-            padding: 0 20px;
-            line-height: 1.4;
-            margin-top: 8px;
-            white-space: pre-wrap;
-            overflow-wrap: break-word;
-          }
-          @media print {
-            body {
-              -webkit-print-color-adjust: exact;
-              color-adjust: exact;
-            }
-          }
-        </style>
-      </head>
-      <body>
-        <div class="qr-code-container">
-          <div class="equipment-name">${equipment.name || "N/A"}</div>
-          <img src="${url}" alt="QR Code" class="qr-code" />
-          <div class="equipment-id">ID: ${equipment.qrId || "N/A"}</div>
-          <div class="equipment-description">${
-            equipment.description
-              ? `Brief Description:\n${equipment.description || ""}`
-              : ""
-          }</div>
-        </div>
-        <script>
-          // Print and close window when loaded
-          window.onload = function() {
-            window.print();
-            // Close after printing is done or canceled
-            setTimeout(() => window.close());
-          };
-        </script>
-      </body>
-      </html>
-    `);
-
-    printWindow.document.close();
-  };
-
-  const cancelDelete = () => {
-    setShowDeleteDialog(false);
-    setPendingDeleteId(null);
-  };
-
-  // Filter equipment by name, make, or model, and by approval status if showPendingOnly is active
-  const filteredEquipment = equipmentDetails.filter((item) => {
-    const term = searchTerm.trim().toLowerCase();
-    if (!term) return true;
-    const nameMatch = item.name.toLowerCase().includes(term);
-    const makeMatch =
-      item.equipmentVehicleInfo?.make?.toLowerCase().includes(term) ?? false;
-    const modelMatch =
-      item.equipmentVehicleInfo?.model?.toLowerCase().includes(term) ?? false;
-    return nameMatch || makeMatch || modelMatch;
-  });
-
   return (
-    <div className="w-full p-4 grid grid-rows-[3rem_2rem_1fr] gap-4">
+    <div className="w-full p-4 grid grid-rows-[3rem_2rem_1fr] gap-5">
       <PageHeaderContainer
         loading={loading}
         headerText="Equipment Management"
@@ -217,22 +70,21 @@ export default function EquipmentPage() {
           rerender();
         }}
       />
-      <div className="h-fit max-h-12  w-full flex flex-row justify-between gap-2 mb-2 ">
-        <div className="flex flex-row w-full gap-2 mb-2">
-          <SearchBarPopover
-            term={searchTerm}
-            handleSearchChange={(e) => setSearchTerm(e.target.value)}
-            placeholder={"Search by name, make, or model..."}
-            textSize="xs"
-            imageSize="10"
-          />
-        </div>
+      <div className="h-10 w-full flex flex-row justify-between gap-4">
+        <SearchBarPopover
+          term={searchTerm}
+          handleSearchChange={(e) => setSearchTerm(e.target.value)}
+          placeholder={"Search by name, make, or model..."}
+          textSize="xs"
+          imageSize="10"
+        />
+
         <div className="flex flex-row justify-end w-full gap-2">
           <Tooltip>
             <TooltipTrigger asChild>
               <Button
                 onClick={() => setCreateEquipmentModal(true)}
-                className="min-w-12 "
+                className="min-w-12 h-full "
               >
                 <img
                   src="/plus-white.svg"
@@ -247,7 +99,7 @@ export default function EquipmentPage() {
             <TooltipTrigger asChild>
               <Button
                 onClick={() => setShowPendingOnly(!showPendingOnly)}
-                className={`relative border-none w-fit min-w-16  px-4 bg-gray-900 hover:bg-gray-800 text-white ${
+                className={`relative border-none w-fit min-w-16 h-full px-4 bg-gray-900 hover:bg-gray-800 text-white ${
                   showPendingOnly ? "ring-2 ring-red-400" : ""
                 }`}
               >
