+++ shift-scan/src/app/(routes)/admins/jobsites/_components/jobsiteTable.tsx
@@ -1,3 +1,4 @@
+"use client";
 import {
   Table,
   TableBody,
@@ -20,6 +21,7 @@ import {
   HoverCardContent,
   HoverCardTrigger,
 } from "@/components/ui/hover-card";
+import { SquareCheck, SquareX } from "lucide-react";
 
 export default function JobsiteTable({
   loading,
@@ -36,31 +38,33 @@ export default function JobsiteTable({
   openHandleQr: (id: string) => void;
   showPendingOnly: boolean;
 }) {
-  const header = [
-    // "ID",
-    "Name",
-    "Description",
-    "Created At",
-    "Active",
-    "Site Address",
-    "Approval Status",
-    "Updated At",
-    "Actions",
-  ];
-
   return (
     <>
       <Table className="w-full mb-10">
         <TableHeader>
           <TableRow>
-            {header.map((h) => (
-              <TableHead
-                key={h}
-                className="text-sm text-center border-r border-gray-200 bg-gray-100"
-              >
-                {h}
-              </TableHead>
-            ))}
+            <TableHead className="text-sm text-center border-r border-gray-200 bg-gray-100">
+              Name & Description
+            </TableHead>
+            <TableHead className="text-sm text-center border-r border-gray-200 bg-gray-100">
+              Created
+            </TableHead>
+            <TableHead className="text-sm text-center border-r border-gray-200 bg-gray-100">
+              Active
+            </TableHead>
+
+            <TableHead className="text-sm text-center border-r border-gray-200 bg-gray-100">
+              Site Address
+            </TableHead>
+            <TableHead className="text-sm text-center border-r border-gray-200 bg-gray-100">
+              Status
+            </TableHead>
+            <TableHead className="text-sm text-center border-r border-gray-200 bg-gray-100">
+              Updated At
+            </TableHead>
+            <TableHead className="text-sm text-center border-r border-gray-200 bg-gray-100">
+              Actions
+            </TableHead>
           </TableRow>
         </TableHeader>
         {loading ? (
@@ -78,17 +82,14 @@ export default function JobsiteTable({
                 <TableCell className="border-r border-gray-200 text-xs text-center">
                   <Skeleton className="h-4 w-3/4 mx-auto" />
                 </TableCell>
-                {/* Description */}
-                <TableCell className="border-r border-gray-200 text-xs text-center">
-                  <Skeleton className="h-4 w-3/4 mx-auto" />
-                </TableCell>
+
                 {/* Created At */}
                 <TableCell className="border-r border-gray-200 text-xs text-center">
                   <Skeleton className="h-4 w-1/2 mx-auto" />
                 </TableCell>
                 {/* Active */}
                 <TableCell className="border-r border-gray-200 text-xs text-center">
-                  <Skeleton className="h-4 w-1/2 mx-auto" />
+                  <Skeleton className="h-4 w-4 mx-auto" />
                 </TableCell>
                 {/* Site Address */}
                 <TableCell className="border-r border-gray-200 text-xs text-center">
@@ -121,22 +122,35 @@ export default function JobsiteTable({
                 className="odd:bg-white even:bg-gray-100 "
                 key={jobsite.id}
               >
-                {/* <TableCell className=" border-r border-gray-200 text-xs text-center">
-                  {jobsite.id || " "}
-                </TableCell> */}
-                <TableCell className=" border-r border-gray-200 text-xs text-center">
-                  {jobsite.name || " "}
+                <TableCell className=" border-r border-gray-200 text-xs text-left">
+                  <div className="flex flex-col gap-1 text-left">
+                    <p>{jobsite.name || " "}</p>
+                    <p className="text-[10px] text-gray-400 italic">
+                      {jobsite.description || "No Description"}
+                    </p>
+                  </div>
                 </TableCell>
-                <TableCell className=" border-r border-gray-200 text-xs text-center">
-                  {jobsite.description || " "}
-                </TableCell>
+
                 <TableCell className=" border-r border-gray-200 text-xs text-center">
                   {jobsite.createdAt
                     ? format(new Date(jobsite.createdAt), "MM/dd/yyyy")
                     : " "}
                 </TableCell>
                 <TableCell className=" border-r border-gray-200 text-xs text-center">
-                  {jobsite.isActive ? "Active" : "Inactive"}
+                  <Tooltip>
+                    <TooltipTrigger asChild>
+                      <div className="flex justify-center">
+                        {jobsite.isActive ? (
+                          <SquareCheck className="h-4 w-4 text-green-500" />
+                        ) : (
+                          <SquareX className="h-4 w-4 text-red-500" />
+                        )}
+                      </div>
+                    </TooltipTrigger>
+                    <TooltipContent>
+                      {jobsite.isActive ? "Active" : "Inactive"}
+                    </TooltipContent>
+                  </Tooltip>
                 </TableCell>
                 <TableCell className=" border-r border-gray-200 text-xs text-center">
                   {jobsite.Address &&
