+++ shift-scan/src/app/(routes)/admins/jobsites/page.tsx
@@ -1,10 +1,9 @@
 "use client";
 import { Button } from "@/components/ui/button";
 import { ScrollArea, ScrollBar } from "@/components/ui/scroll-area";
-import { useSidebar } from "@/components/ui/sidebar";
 import React, { useEffect, useState } from "react";
 import JobsiteTable from "./_components/jobsiteTable";
-import QRCode from "qrcode";
+
 import {
   Dialog,
   DialogContent,
@@ -13,7 +12,6 @@ import {
   DialogHeader,
   DialogTitle,
 } from "@/components/ui/dialog";
-import { deleteJobsite } from "@/actions/AssetActions";
 
 import { Badge } from "@/components/ui/badge";
 import { JobsiteSummary, useJobsiteData } from "./_components/useJobsiteData";
@@ -27,7 +25,6 @@ import {
   PaginationNext,
 } from "@/components/ui/pagination";
 import Spinner from "@/components/(animations)/spinner";
-import ReloadBtnSpinner from "@/components/(animations)/reload-btn-spinner";
 import {
   Tooltip,
   TooltipContent,
@@ -37,11 +34,10 @@ import SearchBarPopover from "../_pages/searchBarPopover";
 import { PageHeaderContainer } from "../_pages/PageHeaderContainer";
 
 export default function JobsitePage() {
-  const { setOpen, open } = useSidebar();
-  const [searchTerm, setSearchTerm] = useState("");
   const {
+    searchTerm,
+    setSearchTerm,
     loading,
-    jobsiteDetails,
     rerender,
     total,
     page,
@@ -50,172 +46,25 @@ export default function JobsitePage() {
     setPageSize,
     showPendingOnly,
     setShowPendingOnly,
+    editJobsiteModal,
+    setEditJobsiteModal,
+    createJobsiteModal,
+    setCreateJobsiteModal,
+    showDeleteDialog,
+    setShowDeleteDialog,
+    pendingEditId,
+    openHandleEdit,
+    openHandleDelete,
+    openHandleQr,
+    confirmDelete,
+    cancelDelete,
+    pendingCount,
+    totalPages,
+    paginatedJobsites,
   } = useJobsiteData();
 
-  // State for modals
-  const [editJobsiteModal, setEditJobsiteModal] = useState(false);
-  const [createJobsiteModal, setCreateJobsiteModal] = useState(false);
-
-  // State for delete confirmation dialog
-  const [showDeleteDialog, setShowDeleteDialog] = useState(false);
-  const [pendingDeleteId, setPendingDeleteId] = useState<string | null>(null);
-  const [pendingEditId, setPendingEditId] = useState<string | null>(null);
-  const [pendingQrId, setPendingQrId] = useState<string | null>(null);
-
-  // Pagination state
-
-  const openHandleEdit = (id: string) => {
-    setPendingEditId(id);
-    setEditJobsiteModal(true);
-  };
-
-  const openHandleDelete = (id: string) => {
-    setPendingDeleteId(id);
-    setShowDeleteDialog(true);
-  };
-
-  const confirmDelete = async () => {
-    if (pendingDeleteId) {
-      await deleteJobsite(pendingDeleteId);
-      setShowDeleteDialog(false);
-      setPendingDeleteId(null);
-      rerender();
-    }
-  };
-
-  const cancelDelete = () => {
-    setShowDeleteDialog(false);
-    setPendingDeleteId(null);
-  };
-
-  // Count all pending items
-  const pendingCount = jobsiteDetails.filter(
-    (item) => item.approvalStatus === "PENDING",
-  ).length;
-
-  // Filter job sites by name or client name and by approval status if showPendingOnly is active
-  const filteredJobsites = jobsiteDetails.filter((item) => {
-    const term = searchTerm.trim().toLowerCase();
-    if (!term || term.length < 3) return true;
-    const jobsiteName = item.name?.toLowerCase() || "";
-    return jobsiteName.includes(term);
-  });
-
-  // Pagination logic
-  const totalJobsites = filteredJobsites.length;
-  const totalPages = Math.ceil(totalJobsites / pageSize);
-  const paginatedJobsites = filteredJobsites.slice(
-    (page - 1) * pageSize,
-    page * pageSize,
-  );
-
-  // Reset to page 1 if search or filter changes
-  useEffect(() => {
-    setPage(1);
-  }, [searchTerm, showPendingOnly]);
-
-  const openHandleQr = (id: string) => {
-    setPendingQrId(id);
-    const jobsite = jobsiteDetails.find((j) => j.id === id);
-    if (jobsite) {
-      printQRCode(jobsite);
-    }
-  };
-
-  const printQRCode = async (jobsite: JobsiteSummary) => {
-    if (!pendingQrId) return;
-    const url = await QRCode.toDataURL(jobsite.qrId || "");
-    // Open a new window for printing
-    const printWindow = window.open("", "_blank");
-    if (!printWindow) {
-      alert("Please allow popups to print the QR code");
-      return;
-    }
-
-    // Write HTML content to the new window
-    printWindow.document.write(`
-      <!DOCTYPE html>
-      <html>
-      <head>
-        <title>Print QR Code - ${jobsite.name || "Jobsite"}</title>
-        <style>
-          body {
-            display: flex;
-            flex-direction: column;
-            align-items: center;
-            justify-content: center;
-            height: 100vh;
-            margin: 0;
-            padding: 20px;
-            box-sizing: border-box;
-            font-family: Arial, sans-serif;
-          }
-          .qr-code-container {
-            text-align: center;
-          }
-          .qr-code {
-            width: 300px;
-            height: 300px;
-            border: 4px solid black;
-            border-radius: 10px;
-            margin-bottom: 20px;
-          }
-          .equipment-name {
-            font-size: 24px;
-            font-weight: bold;
-            margin-bottom: 10px;
-          }
-          .equipment-id {
-            font-size: 16px;
-            color: #555;
-            margin-bottom: 8px;
-          }
-          .equipment-description {
-            font-size: 16px;
-            color: #555;
-            max-width: 350px;
-            padding: 0 20px;
-            line-height: 1.4;
-            margin-top: 8px;
-            white-space: pre-wrap;
-            overflow-wrap: break-word;
-          }
-          @media print {
-            body {
-              -webkit-print-color-adjust: exact;
-              color-adjust: exact;
-            }
-          }
-        </style>
-      </head>
-      <body>
-        <div class="qr-code-container">
-          <div class="equipment-name">${jobsite.name || "N/A"}</div>
-          <img src="${url}" alt="QR Code" class="qr-code" />
-          <div class="equipment-id">ID: ${jobsite.qrId || "N/A"}</div>
-          <div class="equipment-description">${
-            jobsite.description
-              ? `Brief Description:\n${jobsite.description || ""}`
-              : ""
-          }</div>
-        </div>
-        <script>
-          // Print and close window when loaded
-          window.onload = function() {
-            window.print();
-            // Close after printing is done or canceled
-            setTimeout(() => window.close());
-          };
-        </script>
-      </body>
-      </html>
-    `);
-
-    printWindow.document.close();
-  };
-
   return (
-    <div className="w-full p-4 grid grid-rows-[3rem_2rem_1fr] gap-4">
+    <div className="w-full p-4 grid grid-rows-[3rem_2rem_1fr] gap-5">
       <PageHeaderContainer
         loading={loading}
         headerText="Jobsite Management"
@@ -224,23 +73,22 @@ export default function JobsitePage() {
           rerender();
         }}
       />
-      <div className="h-fit max-h-12  w-full flex flex-row justify-between gap-2 mb-2 ">
-        <div className="flex flex-row w-full gap-4 mb-2">
-          <SearchBarPopover
-            term={searchTerm}
-            handleSearchChange={(e) => setSearchTerm(e.target.value)}
-            placeholder={"Search by name or client..."}
-            textSize="xs"
-            imageSize="10"
-          />
-        </div>
+      <div className="h-10 w-full flex flex-row justify-between">
+        <SearchBarPopover
+          term={searchTerm}
+          handleSearchChange={(e) => setSearchTerm(e.target.value)}
+          placeholder={"Search by name or client..."}
+          textSize="xs"
+          imageSize="10"
+        />
+
         <div className="flex flex-row justify-end w-full gap-2">
           <Tooltip>
             <TooltipTrigger asChild>
               <Button
                 size={"icon"}
                 onClick={() => setCreateJobsiteModal(true)}
-                className="min-w-12"
+                className="min-w-12 h-full"
               >
                 <img
                   src="/plus-white.svg"
@@ -258,7 +106,7 @@ export default function JobsitePage() {
               <Button
                 size={"icon"}
                 onClick={() => setShowPendingOnly(!showPendingOnly)}
-                className={`relative border-none min-w-16  bg-gray-900 hover:bg-gray-800 text-white ${
+                className={`relative border-none min-w-16 h-full  bg-gray-900 hover:bg-gray-800 text-white ${
                   showPendingOnly ? "ring-2 ring-red-400" : ""
                 }`}
               >
